% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMA_model-internals.R
\name{ARMA_companion_matrix}
\alias{ARMA_companion_matrix}
\title{Construct the companion matrix of an ARMA model}
\usage{
ARMA_companion_matrix(phi, theta)
}
\arguments{
\item{phi}{Numeric vector with length \code{p}, AR parameters.}

\item{theta}{Numeric vector with length \code{q}, MA parameters.}
}
\value{
A square matrix with dimension (p+q)
}
\description{
Construct the companion matrix of an ARMA model
}
\note{
Version 1.0.0.
}
\examples{
# AR(1) / MA(1) ~ No companion
ARMA_companion_matrix(c(0.4))
ARMA_companion_matrix(theta = c(0.4))
# Only AR
ARMA_companion_matrix(c(0.4, 0.3, 0.1))
# Only MA
ARMA_companion_matrix(theta = c(0.4, 0.3, 0.1))
# ARMA
ARMA_companion_matrix(c(0.4, 0.2), c(0.3))
ARMA_companion_matrix(c(0.1, 0.02, 0.01), c(0.3, 0.1, 0.05))
}
\keyword{ARMA}
