% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMA_model-internals.R
\name{ARMA_forecast}
\alias{ARMA_forecast}
\title{Fast ARMA state-space h-step forecast and weights (C/BLAS)}
\usage{
ARMA_forecast(h, X0, A, b, intercept = 0)
}
\arguments{
\item{h}{Integer, steps ahead.}

\item{X0}{Numeric vector of length p+q (state).}

\item{A}{Numeric (p+q) x (p+q) companion matrix.}

\item{b}{Numeric vector length p+q (shocks selector).}

\item{intercept}{Scalar intercept (0 if none).}
}
\description{
Fast ARMA state-space h-step forecast and weights (C/BLAS)
}
\examples{
h <- 1000
X0 <- c(0.2, 0.1)
A <- ARMA_companion_matrix(0.2, 0.1)
b <- ARMA_vector_b(1,1)
ARMA_forecast(h, X0, A, b, intercept = 0)

}
