% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMA_model-internals.R
\name{ARMA_loss_CSS}
\alias{ARMA_loss_CSS}
\title{Sum of the errors with contraint parameters}
\usage{
ARMA_loss_CSS(params, p = 0, q = 0, y)
}
\description{
Sum of the errors with contraint parameters
}
