% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMA_model-internals.R
\name{ARMA_loss_logLik}
\alias{ARMA_loss_logLik}
\title{Log-likelihood with contraint parameters}
\usage{
ARMA_loss_logLik(params, p = 0, q = 0, y, per_obs = FALSE)
}
\description{
Log-likelihood with contraint parameters
}
