% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMA_model-R6.R
\name{ARMA_modelR6}
\alias{ARMA_modelR6}
\title{R6 class for ARMA(p, q) model}
\description{
R6 class for ARMA(p, q) model

R6 class for ARMA(p, q) model
}
\note{
Version 1.0.1
}
\seealso{
\code{\link[stats:arima]{stats::arima()}} which is wrapped in the method \code{fit}.
}
\keyword{ARMA}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{model}}{An object with the fitted ARMA model from the function \code{\link[stats:arima]{stats::arima()}}.}

\item{\code{intercept}}{Numeric named scalar, intercept of the model.}

\item{\code{phi}}{Numeric named vector, AR parameters.}

\item{\code{theta}}{Numeric named vector, MA parameters.}

\item{\code{order}}{Numeric named vector, ARMA order. The first element is the AR order, while the second the MA order.}

\item{\code{coefficients}}{Numeric named vector, intercept and ARMA parameters.}

\item{\code{std.errors}}{Numeric named vector, std.errors of the intercept and ARMA parameters.}

\item{\code{sigma2}}{Numeric scalar, std.errors of the residuals.}

\item{\code{A}}{Numeric matrix, companion matrix to govern the transition between two time steps.  See the function \code{\link[=ARMA_companion_matrix]{ARMA_companion_matrix()}}.}

\item{\code{b}}{Numeric vector, unitary vector for the residuals. See the function \code{\link[=ARMA_vector_b]{ARMA_vector_b()}}.}

\item{\code{tidy}}{Tibble with estimated parameters and relative std. errors.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ARMA_modelR6-new}{\code{ARMA_modelR6$new()}}
\item \href{#method-ARMA_modelR6-fit}{\code{ARMA_modelR6$fit()}}
\item \href{#method-ARMA_modelR6-filter}{\code{ARMA_modelR6$filter()}}
\item \href{#method-ARMA_modelR6-next_step}{\code{ARMA_modelR6$next_step()}}
\item \href{#method-ARMA_modelR6-expectation}{\code{ARMA_modelR6$expectation()}}
\item \href{#method-ARMA_modelR6-variance}{\code{ARMA_modelR6$variance()}}
\item \href{#method-ARMA_modelR6-update}{\code{ARMA_modelR6$update()}}
\item \href{#method-ARMA_modelR6-update_std.errors}{\code{ARMA_modelR6$update_std.errors()}}
\item \href{#method-ARMA_modelR6-update_sigma2}{\code{ARMA_modelR6$update_sigma2()}}
\item \href{#method-ARMA_modelR6-print}{\code{ARMA_modelR6$print()}}
\item \href{#method-ARMA_modelR6-clone}{\code{ARMA_modelR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARMA_modelR6-new"></a>}}
\if{latex}{\out{\hypertarget{method-ARMA_modelR6-new}{}}}
\subsection{Method \code{new()}}{
Initialize an ARMA model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARMA_modelR6$new(arOrder = 1, maOrder = 1, include.intercept = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{arOrder}}{Numeric scalar, order for Autoregressive component.}

\item{\code{maOrder}}{Numeric scalar, order for Moving-Average component.}

\item{\code{include.intercept}}{Logical. When \code{TRUE} the intercept will be included. The default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARMA_modelR6-fit"></a>}}
\if{latex}{\out{\hypertarget{method-ARMA_modelR6-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the ARMA model with \code{arima} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARMA_modelR6$fit(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Numeric vector, time series to fit.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARMA_modelR6-filter"></a>}}
\if{latex}{\out{\hypertarget{method-ARMA_modelR6-filter}{}}}
\subsection{Method \code{filter()}}{
Filter the time-series and compute fitted values and residuals.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARMA_modelR6$filter(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Numeric vector, time series to filter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARMA_modelR6-next_step"></a>}}
\if{latex}{\out{\hypertarget{method-ARMA_modelR6-next_step}{}}}
\subsection{Method \code{next_step()}}{
Next step function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARMA_modelR6$next_step(x, n.ahead = 1, eps = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Numeric vector, state vector with past observations and residuals.}

\item{\code{n.ahead}}{Numeric scalar, forecasted steps ahead.}

\item{\code{eps}}{Numeric vector, optional realized residuals.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARMA_modelR6-expectation"></a>}}
\if{latex}{\out{\hypertarget{method-ARMA_modelR6-expectation}{}}}
\subsection{Method \code{expectation()}}{
Forecast expected value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARMA_modelR6$expectation(h = 1, X0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{h}}{Numeric scalar, number of steps ahead.}

\item{\code{X0}}{Numeric vector with length \code{p + q}, state vector of past values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARMA_modelR6-variance"></a>}}
\if{latex}{\out{\hypertarget{method-ARMA_modelR6-variance}{}}}
\subsection{Method \code{variance()}}{
Forecast variance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARMA_modelR6$variance(h = 1, sigma2 = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{h}}{Numeric scalar, number of steps ahead.}

\item{\code{sigma2}}{Numeric scalar, std. deviation of the residuals.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARMA_modelR6-update"></a>}}
\if{latex}{\out{\hypertarget{method-ARMA_modelR6-update}{}}}
\subsection{Method \code{update()}}{
Update the model's parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARMA_modelR6$update(coefficients)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coefficients}}{Numeric named vector, model's coefficients. If missing nothing will be updated.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARMA_modelR6-update_std.errors"></a>}}
\if{latex}{\out{\hypertarget{method-ARMA_modelR6-update_std.errors}{}}}
\subsection{Method \code{update_std.errors()}}{
Update the standard errors of the parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARMA_modelR6$update_std.errors(std.errors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{std.errors}}{Numeric named vector, parameters' standard errors. If missing nothing will be updated.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARMA_modelR6-update_sigma2"></a>}}
\if{latex}{\out{\hypertarget{method-ARMA_modelR6-update_sigma2}{}}}
\subsection{Method \code{update_sigma2()}}{
Update the variance of the residuals.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARMA_modelR6$update_sigma2(sigma2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sigma2}}{Numeric scalar, variance of the residuals.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARMA_modelR6-print"></a>}}
\if{latex}{\out{\hypertarget{method-ARMA_modelR6-print}{}}}
\subsection{Method \code{print()}}{
Print method for \code{AR_modelR6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARMA_modelR6$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARMA_modelR6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ARMA_modelR6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARMA_modelR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
