% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarScenario.R
\name{as_solarScenario}
\alias{as_solarScenario}
\title{Extract and structure simulations from a \code{solarScenarioSpec}}
\usage{
as_solarScenario(simSpec, by)
}
\arguments{
\item{simSpec}{object with the class \code{solarScenario_spec}. See the function \code{\link{solarScenario_spec}} for details.}

\item{by}{Optional character. Represent the steps used for multiple scenarios.}
}
\description{
Extract and structure simulations from a \code{solarScenarioSpec}
}
\note{
Version 1.0.0.
}
\keyword{solarScenario}
