% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundTransform-R6.R
\name{boundTransform}
\alias{boundTransform}
\title{Bounded transformation functions}
\description{
Bounded transformation functions

Bounded transformation functions
}
\note{
Version 1.0.0.
}
\examples{
st <- boundTransform$new()

}
\keyword{utils}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{epsilon}}{Numeric, \eqn{\epsilon} transformation parameter.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{alpha}}{Numeric, \eqn{\alpha} transformation parameter.}

\item{\code{beta}}{Numeric, \eqn{\beta} transformation parameter.}

\item{\code{monotonicity}}{Character, type of monotonicity of \eqn{g}.}

\item{\code{link}}{Character, name of the link function \eqn{g}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-boundTransform-new}{\code{boundTransform$new()}}
\item \href{#method-boundTransform-set_transform}{\code{boundTransform$set_transform()}}
\item \href{#method-boundTransform-X_prime}{\code{boundTransform$X_prime()}}
\item \href{#method-boundTransform-iX_prime}{\code{boundTransform$iX_prime()}}
\item \href{#method-boundTransform-Y}{\code{boundTransform$Y()}}
\item \href{#method-boundTransform-iY}{\code{boundTransform$iY()}}
\item \href{#method-boundTransform-g}{\code{boundTransform$g()}}
\item \href{#method-boundTransform-ig}{\code{boundTransform$ig()}}
\item \href{#method-boundTransform-g_prime}{\code{boundTransform$g_prime()}}
\item \href{#method-boundTransform-fit}{\code{boundTransform$fit()}}
\item \href{#method-boundTransform-bounds}{\code{boundTransform$bounds()}}
\item \href{#method-boundTransform-update}{\code{boundTransform$update()}}
\item \href{#method-boundTransform-print}{\code{boundTransform$print()}}
\item \href{#method-boundTransform-clone}{\code{boundTransform$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-boundTransform-new"></a>}}
\if{latex}{\out{\hypertarget{method-boundTransform-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{solarTransform} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{boundTransform$new(alpha = 0, beta = 1, link = "invgumbel")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{Numeric, \eqn{\alpha} transformation parameter.}

\item{\code{beta}}{Numeric, \eqn{\beta} transformation parameter.}

\item{\code{link}}{Character, link function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-boundTransform-set_transform"></a>}}
\if{latex}{\out{\hypertarget{method-boundTransform-set_transform}{}}}
\subsection{Method \code{set_transform()}}{
Set the transform function \eqn{g}, \eqn{g^{-1}}, \eqn{g^{\prime}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{boundTransform$set_transform(link)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{link}}{Character, link function. Valid links are \code{"invgumbel"}, \code{"gumbel"}, \code{"logis"}, \code{"norm"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-boundTransform-X_prime"></a>}}
\if{latex}{\out{\hypertarget{method-boundTransform-X_prime}{}}}
\subsection{Method \code{X_prime()}}{
Map the risk-driver \eqn{X_t} into the normalized risk-driver \eqn{X_t^{\prime}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{boundTransform$X_prime(Xt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Xt}}{Numeric, risk driver in \eqn{ X_t \in (\alpha, \alpha+\beta)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{\text{X}^{\prime}(X_t) = \frac{X_t - \alpha}{\beta}}
}

\subsection{Returns}{
Numeric, normalized risk driver \eqn{X_t^{\prime} \in (0, 1)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-boundTransform-iX_prime"></a>}}
\if{latex}{\out{\hypertarget{method-boundTransform-iX_prime}{}}}
\subsection{Method \code{iX_prime()}}{
Map the normalized variable \eqn{X_t^{\prime}} to the risk driver \eqn{X_t}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{boundTransform$iX_prime(Xt_prime)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Xt_prime}}{Numeric, normalized risk driver \eqn{X_t^{\prime} \in (0, 1)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{\text{iX}^{\prime}(X_t^{\prime}) = \alpha + \beta \cdot X_t^{\prime}}
}

\subsection{Returns}{
Numeric, risk driver in \eqn{ X_t \in (\alpha, \alpha+\beta)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-boundTransform-Y"></a>}}
\if{latex}{\out{\hypertarget{method-boundTransform-Y}{}}}
\subsection{Method \code{Y()}}{
Map the normalized risk driver \eqn{X_t^{\prime}} in the transformed variable \eqn{Y_t}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{boundTransform$Y(Xt_prime)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Xt_prime}}{Numeric, normalized risk driver \eqn{X_t^{\prime} \in (0, 1)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{\text{Y}(X_t^{\prime}) = g(X_t^{\prime})}
}

\subsection{Returns}{
Numeric, transformed variable \eqn{Y_t \in (-\infty, \infty)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-boundTransform-iY"></a>}}
\if{latex}{\out{\hypertarget{method-boundTransform-iY}{}}}
\subsection{Method \code{iY()}}{
Map the transformed variable \eqn{Y_t} into the normalized risk driver \eqn{X_t^{\prime}}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{boundTransform$iY(Yt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Yt}}{Numeric, transformed variable \eqn{Y_t \in (-\infty, \infty)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{\text{iY}(Y_t) = g^{-1}(Y_t)}
}

\subsection{Returns}{
Numeric, normalized risk driver \eqn{X_t^{\prime} \in (0, 1)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-boundTransform-g"></a>}}
\if{latex}{\out{\hypertarget{method-boundTransform-g}{}}}
\subsection{Method \code{g()}}{
Link function to map \eqn{X_t^{\prime}} to \eqn{Y_t}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{boundTransform$g(X_prime)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_prime}}{Numeric, normalized risk driver \eqn{X_t^{\prime} \in (0, 1)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Numeric, transformed variable \eqn{Y_t \in (-\infty, \infty)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-boundTransform-ig"></a>}}
\if{latex}{\out{\hypertarget{method-boundTransform-ig}{}}}
\subsection{Method \code{ig()}}{
Inverse of the function to map \eqn{Y_t} to \eqn{X^{\prime}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{boundTransform$ig(Yt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Yt}}{Numeric, transformed variable \eqn{Y_t \in (-\infty, \infty)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Numeric, normalized risk driver \eqn{X_t^{\prime} \in (0, 1)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-boundTransform-g_prime"></a>}}
\if{latex}{\out{\hypertarget{method-boundTransform-g_prime}{}}}
\subsection{Method \code{g_prime()}}{
First derivative of the function \eqn{g}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{boundTransform$g_prime(X_prime)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_prime}}{Numeric, normalized risk driver \eqn{X_t^{\prime} \in (0, 1)}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-boundTransform-fit"></a>}}
\if{latex}{\out{\hypertarget{method-boundTransform-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the best parameters \eqn{\alpha} and \eqn{\beta} from a given time series
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{boundTransform$fit(x, epsilon = 0.01, min_pos = 1, max_pos = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{time series of solar risk drivers in \eqn{(0, 1)}.}

\item{\code{epsilon}}{Numeric}

\item{\code{min_pos}}{Integer, position of the minimum. For example when \code{2} the minimum is the second lowest value.}

\item{\code{max_pos}}{Integer, position of the maximum. For example when \code{3} the maximum is the third greatest value.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Return a list that contains:
\describe{
\item{alpha}{Numeric, \eqn{\alpha} transformation parameter.}
\item{beta}{Numeric, \eqn{\beta} transformation parameter.}
\item{epsilon}{Numeric, threshold used for fitting.}
\item{Xt_min}{Numeric, minimum value of the time series.}
\item{Xt_min}{Numeric, maximum value of the time series.}
}
}

\subsection{Returns}{
A named list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-boundTransform-bounds"></a>}}
\if{latex}{\out{\hypertarget{method-boundTransform-bounds}{}}}
\subsection{Method \code{bounds()}}{
Compute the bounds for the transformed variables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{boundTransform$bounds(target = "Xt")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target}}{target variable. Available choices are:
\describe{
\item{\code{"Xt"}}{Solar risk driver, the bounds returned are \eqn{[\alpha, \alpha + \beta]}.}
\item{\code{"Kt"}}{Clearness index, the bounds returned are \eqn{[1-\alpha-\beta, 1-\alpha]}.}
\item{\code{"Yt"}}{Solar transform, the bounds returned are \eqn{[-\infty, \infty]}.}
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector where the first element is the lower bound and the second the upper bound.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-boundTransform-update"></a>}}
\if{latex}{\out{\hypertarget{method-boundTransform-update}{}}}
\subsection{Method \code{update()}}{
Update the transformation parameters \eqn{\alpha} and \eqn{\beta}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{boundTransform$update(alpha, beta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{Numeric, transformation parameter.}

\item{\code{beta}}{Numeric, transformation parameter.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Update the slots \verb{$alpha} and \verb{$beta}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-boundTransform-print"></a>}}
\if{latex}{\out{\hypertarget{method-boundTransform-print}{}}}
\subsection{Method \code{print()}}{
Print method for the class \code{solarTransform}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{boundTransform$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-boundTransform-clone"></a>}}
\if{latex}{\out{\hypertarget{method-boundTransform-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{boundTransform$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
