% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalClearsky-R6-internals.R
\name{clearsky_optimizer}
\alias{clearsky_optimizer}
\title{Optimizer for clear sky model with restricted least squares (RLS).}
\usage{
clearsky_optimizer(seasonal_model_Ct, newdata, ntol = 0)
}
\arguments{
\item{seasonal_model_Ct}{An object of the class \code{seasonalClearsky}. See the function \code{\link{seasonalClearsky}} for more details.}

\item{newdata}{A data frame to input to the method \code{seasonal_model_Ct$predict()}.}

\item{ntol}{Integer scalar. Tolerance for the maximum number of violations admitted of the condition \code{clearsky > GHI}. Default is \code{0}.}
}
\description{
Optimizer for clear sky model with restricted least squares (RLS).
}
\note{
Version 1.0.0
}
\keyword{clearsky}
