% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalClearsky-R6-internals.R
\name{clearsky_outliers}
\alias{clearsky_outliers}
\title{Impute clear sky outliers}
\usage{
clearsky_outliers(x, Ct, date, threshold = 1e-04, quiet = FALSE)
}
\arguments{
\item{x}{Numeric vector, time series of solar radiation.}

\item{Ct}{Numeric vector, time series of fitted clear sky radiation.}

\item{date}{Character or Date vector, time series of dates used to precisely impute solar radiation according to the realized values in the same day of the year.}

\item{threshold}{Numeric, scalar, threshold value used for imputation. Default is \code{0.0001}.}

\item{quiet}{Logical.}
}
\description{
Detect and impute outliers with respect to a maximum level of radiation (Ct)
}
\details{
The function will detect the observations for which \code{x > Ct}, \code{x < 0} or \code{is.na(x)}. Then, if
\describe{
\item{\code{x < 0}}{If a value is below 0 for a day it will be imputed to be equal to \code{min(x)} for that day}.
\item{\code{x > Ct}}{If a value is above the maximum clear sky Ct it will be imputed to be \code{Ct*(1-threshold)}}.
\item{\code{is.na(x)}}{If a value is NA it will be imputed to be the average \code{mean(x)} for that day.}.
}
}
\note{
Version 1.0.0
}
\examples{
clearsky_outliers(c(1,2,3), 2)
}
\keyword{clearsky}
