% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalClearsky-R6-internals.R
\name{control_seasonalClearsky}
\alias{control_seasonalClearsky}
\title{Control parameters for a \code{seasonalClearsky} object}
\usage{
control_seasonalClearsky(
  order = 1,
  order_H0 = 1,
  period = 365,
  include.intercept = TRUE,
  include.trend = FALSE,
  delta0 = 1.4,
  lower = 0,
  upper = 3,
  by = 0.001,
  ntol = 0,
  quiet = FALSE
)
}
\arguments{
\item{order}{Integer scalar, number of combinations of sines and cosines.}

\item{period}{Integer scalar, seasonality period. The default is 365.}

\item{include.intercept}{Logical, when \code{TRUE}, the default, the intercept will be included in the clear sky model.}

\item{delta0}{Numeric scalar, initial value for the optimization. The estimated clear sky is increased by delta0.}

\item{ntol}{Integer scalar. Tolerance for the maximum number of violations admitted of the condition \code{clearsky > GHI}. Default is \code{0}.}

\item{quiet}{Logical, when \code{FALSE}, the default, the functions displays warning or messages.}
}
\value{
Named list of control parameters.
}
\description{
Control parameters for a \code{seasonalClearsky} object
}
\details{
The parameters \code{ntol}, \code{lower}, \code{upper} and \code{by} are used exclusively in \code{\link{clearsky_optimizer}}.
}
\note{
Version 1.0.0
}
\examples{
control = control_seasonalClearsky()
}
\keyword{control}
