% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarHedging.R
\name{control_solarHedging}
\alias{control_solarHedging}
\title{Control parameters for solar hedging}
\usage{
control_solarHedging(
  n_panels = 1,
  efficiency = 1,
  PUN = 1,
  tick = 1,
  n_contracts = 1
)
}
\arguments{
\item{n_panels}{Numeric scalar, number of meters squared of solar panels.}

\item{efficiency}{Numeric scalar, average electricity produced with 1 \eqn{m^2} of solar panels given 1 kWh/\eqn{m^2} of solar radiation received.}

\item{PUN}{Numeric scalar, fixed electricity price at which the produced energy is sold.}

\item{tick}{Numeric scalar, tick for the monetary conversion of the payoff of a solar derivative from kWh/\eqn{m^2} to Euros.}

\item{n_contract}{Numeric scalar, number of solar derivative contracts bought.}
}
\description{
Control parameters for solar hedging
}
\note{
Version 1.0.0.
}
\keyword{control}
