% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption-internals.R
\name{control_solarOption}
\alias{control_solarOption}
\title{Control parameters for a solar option}
\usage{
control_solarOption(
  nyears = c(2005, 2025),
  K = 0,
  leap_year = FALSE,
  nsim = 200,
  ci = 0.05,
  seed = 1
)
}
\arguments{
\item{nyears}{numeric vector. Interval of years considered. The first element will be the minimum and the second the maximum years used in
the computation of the fair payoff.}

\item{K}{numeric, level for the strike with respect to the seasonal mean. The seasonal mean is multiplied by \code{exp(K)}.}

\item{leap_year}{logical, when \code{FALSE}, the default, the year will be considered of 365 days, otherwise 366.}

\item{nsim}{integer, number of simulations used to bootstrap the premium's bounds. See \code{\link{solarOption_historical_bootstrap}}.}

\item{ci}{numeric, confidence interval for bootstrapping. See \code{\link{solarOption_historical_bootstrap}}.}

\item{seed}{integer, random seed for reproducibility. See \code{\link{solarOption_historical_bootstrap}}.}
}
\description{
Control parameters for a solar option
}
\note{
Version 1.0.0.
}
\examples{
control_options <- control_solarOption()

}
\keyword{control}
