% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desscher.R
\name{desscher}
\alias{desscher}
\alias{pesscher}
\title{Esscher-distorted density and distribution}
\usage{
desscher(pdf, theta = 0, lower = -Inf, upper = Inf)

pesscher(pdf, theta = 0, lower = -Inf, upper = Inf)
}
\arguments{
\item{pdf}{Function, density function to distort.}

\item{theta}{Numeric, distortion parameter.}

\item{lower, upper}{Numeric, lower and upper bounds for integration, i.e. the bounds of the pdf.}
}
\value{
A function of x.
}
\description{
Given a function of \code{x}, i.e. \eqn{f_X(x)}, compute its Esscher transform and return again a function of \code{x}.
}
\details{
Given a pdf \eqn{f_X(x)} the function computes its Esscher transform, i.e.

\deqn{\mathcal{E}_{\theta}\{f_X(x)\} = \frac{e^{\theta x} f_X(x)}{\int_{-\infty}^{\infty} e^{\theta x} f_X(x) dx}}

Version 1.0.0.
}
\examples{
# Grid of points
grid <- seq(-3, 3, 0.1)
# Density function of x
pdf <- function(x) dnorm(x, mean = 0)
# Esscher density (no transform)
esscher_pdf <- desscher(pdf, theta = 0)
pdf(grid) - esscher_pdf(grid)
# Esscher density (transform)
esscher_pdf_1 <- function(x) dnorm(x, mean = -0.1)
esscher_pdf_2 <- desscher(pdf, theta = -0.1)
esscher_pdf_1(grid) - esscher_pdf_2(grid)

# Esscher Distribution (transform)
esscher_cdf <- pesscher(pdf, theta = -0.1)
plot(esscher_cdf(grid))

}
\keyword{distributions}
