% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{detect_season}
\alias{detect_season}
\title{Detect the season}
\usage{
detect_season(x, invert = FALSE)
}
\arguments{
\item{x}{vector of dates in the format \code{YYYY-MM-DD}.}

\item{invert}{logica, when \code{TRUE} the seasons will be inverted.}
}
\value{
a character vector containing the correspondent season. Can be \code{spring}, \code{summer}, \code{autumn}, \code{winter}.
}
\description{
Detect the season from a vector of dates
}
\examples{
detect_season("2040-01-31")
detect_season(c("2040-01-31", "2023-04-01", "2015-09-02"))

}
