% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmvmixnorm.R
\name{dmvmixnorm}
\alias{dmvmixnorm}
\alias{pmvmixnorm}
\alias{qmvmixnorm}
\title{Multivariate Gaussian mixture random variable}
\usage{
dmvmixnorm(
  x,
  means = matrix(0, 2, 2),
  sigma2 = matrix(1, 2, 2),
  p = rep(1/2, 2),
  rho = c(0, 0),
  log = FALSE
)

pmvmixnorm(
  x,
  means = matrix(0, 2, 2),
  sigma2 = matrix(1, 2, 2),
  p = rep(1/2, 2),
  rho = c(0, 0),
  lower = -Inf,
  log.p = FALSE
)

qmvmixnorm(
  x,
  means = matrix(0, 2, 2),
  sigma2 = matrix(1, 2, 2),
  p = rep(1/2, 2),
  rho = c(0, 0),
  log.p = FALSE
)
}
\description{
Multivariate Gaussian mixture density, distribution, quantile and random generator.
}
\details{
Version 1.0.0.
}
\examples{
# Means components
mean_1 = c(-1.8,-0.4)
mean_2 = c(0.6, 0.5)
# Dimension of the random variable
j = length(mean_1)
# Matrix of means
means = matrix(c(mean_1, mean_2), j,j, byrow = TRUE)

# Variance components
var_1 = c(1,1.4)
var_2 = c(1.3, 1.2)
# Matrix of variances
sigma2 = matrix(c(var_1, var_2), j,j, byrow = TRUE)

# Correlations
rho <- c(rho_1 = 0.2, rho_2 = 0.3)

# Probability for each component
p <- c(0.4, 0.6)

x <- matrix(c(0.1,-0.1), nrow = 1)
dmvmixnorm(x, means, sigma2, p, rho)
pmvmixnorm(x, means, sigma2, p, rho)
qmvmixnorm(0.35, means, sigma2, p, rho)
}
\keyword{distributions}
