% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmvsolarGHI.R
\name{dmvsolarGHI}
\alias{dmvsolarGHI}
\title{Bivariate PDF GHI}
\usage{
dmvsolarGHI(x, Ct, alpha, beta, joint_pdf_Yt)
}
\arguments{
\item{x}{vector of quantiles.}

\item{Ct}{clear sky radiation}

\item{alpha}{parameters \code{alpha > 0}.}

\item{beta}{parameters \code{beta > 0} and \code{alpha + beta < 1}.}

\item{joint_pdf_Yt}{joint density of Y1_t, Y2_t.}
}
\description{
Bivariate PDF GHI
}
\details{
Version 1.0.0.
}
\keyword{distributions}
