% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsugeno.R
\name{dsugeno}
\alias{dsugeno}
\alias{psugeno}
\title{Sugeno-distorted density and distribution}
\usage{
dsugeno(pdf, cdf, lambda = 0)

psugeno(cdf, lambda = 0)
}
\arguments{
\item{pdf}{Function, density function.}

\item{cdf}{Function, distribution function.}

\item{lambda}{Numeric, distortion parameter.}
}
\description{
Compute the Sugeno-distorted distribution for a given distribution or density.
}
\details{
Version 1.0.0.
}
\examples{
# Distribution and density
cdf <- function(x) pnorm(x)
pdf <- function(x) dnorm(x)
x <- seq(-4, 4, 0.01)
plot(x, psugeno(cdf, lambda = -0.2)(x))
plot(x, dsugeno(pdf, cdf, lambda = -0.2)(x))

}
\keyword{distributions}
