% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARCH_model-R6.R
\name{GARCH_modelR6}
\alias{GARCH_modelR6}
\title{Implementation of \code{rugarch} methods for a GARCH(p,q) as R6 class}
\description{
Implementation of \code{rugarch} methods for a GARCH(p,q) as R6 class

Implementation of \code{rugarch} methods for a GARCH(p,q) as R6 class
}
\note{
Version 1.0.0
}
\keyword{GARCH}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{spec}}{A \code{rugarch} object containing the model's specification.}

\item{\code{omega}}{Numeric scalar, intercept parameter.}

\item{\code{alpha}}{Numeric vector, ARCH parameters.}

\item{\code{beta}}{Numeric vector, GARCH parameters.}

\item{\code{order}}{Numeric scalar, model's order.}

\item{\code{coefficients}}{Numeric vector, model's coefficients.}

\item{\code{vol}}{Numeric scalar, long-term unconditional std. deviation.}

\item{\code{loglik}}{model log-likelihood}

\item{\code{tidy}}{Tibble with estimated parameters and relative std. errors.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GARCH_modelR6-new}{\code{GARCH_modelR6$new()}}
\item \href{#method-GARCH_modelR6-fit}{\code{GARCH_modelR6$fit()}}
\item \href{#method-GARCH_modelR6-filter}{\code{GARCH_modelR6$filter()}}
\item \href{#method-GARCH_modelR6-logLik}{\code{GARCH_modelR6$logLik()}}
\item \href{#method-GARCH_modelR6-update}{\code{GARCH_modelR6$update()}}
\item \href{#method-GARCH_modelR6-update_hessian}{\code{GARCH_modelR6$update_hessian()}}
\item \href{#method-GARCH_modelR6-update_std.errors}{\code{GARCH_modelR6$update_std.errors()}}
\item \href{#method-GARCH_modelR6-next_step}{\code{GARCH_modelR6$next_step()}}
\item \href{#method-GARCH_modelR6-print}{\code{GARCH_modelR6$print()}}
\item \href{#method-GARCH_modelR6-clone}{\code{GARCH_modelR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GARCH_modelR6-new"></a>}}
\if{latex}{\out{\hypertarget{method-GARCH_modelR6-new}{}}}
\subsection{Method \code{new()}}{
Initialize a GARCH model with \code{rugarch} specification
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GARCH_modelR6$new(spec, x, weights, sigma20)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{spec}}{GARCH specification from \code{ugarchspec}.}

\item{\code{x}}{Numeric, vector. Time series to be fitted.}

\item{\code{weights}}{Numeric, vector. Optional custom weights.}

\item{\code{sigma20}}{Numeric scalar. Target unconditional variance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GARCH_modelR6-fit"></a>}}
\if{latex}{\out{\hypertarget{method-GARCH_modelR6-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the GARCH model with \code{rugarch} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GARCH_modelR6$fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GARCH_modelR6-filter"></a>}}
\if{latex}{\out{\hypertarget{method-GARCH_modelR6-filter}{}}}
\subsection{Method \code{filter()}}{
Filter method from \code{rugarch} package to compute GARCH variance, residuals and log-likelihoods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GARCH_modelR6$filter(x, coefficients, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Numeric, vector. Time series to be filtered.}

\item{\code{coefficients}}{Numeric, named vector. Model's coefficients. When missing will be used the fitted parameters.}

\item{\code{...}}{Other arguments passed to \code{ugarchfilter} function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GARCH_modelR6-logLik"></a>}}
\if{latex}{\out{\hypertarget{method-GARCH_modelR6-logLik}{}}}
\subsection{Method \code{logLik()}}{
Log-likelihoods function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GARCH_modelR6$logLik(coefficients, x, weights, update = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coefficients}}{Numeric, named vector. Model's coefficients. When missing will be used the fitted parameters.}

\item{\code{x}}{Numeric, vector. Time series used to compute log-likelihoods.}

\item{\code{weights}}{Numeric, vector. Optional custom weights.}

\item{\code{update}}{Logical. When true the internal log-likelihood will be updated.}

\item{\code{...}}{Other arguments passed to \code{ugarchfilter} function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GARCH_modelR6-update"></a>}}
\if{latex}{\out{\hypertarget{method-GARCH_modelR6-update}{}}}
\subsection{Method \code{update()}}{
Update the coefficients of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GARCH_modelR6$update(coefficients)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coefficients}}{Numeric, named vector. Model's coefficients.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GARCH_modelR6-update_hessian"></a>}}
\if{latex}{\out{\hypertarget{method-GARCH_modelR6-update_hessian}{}}}
\subsection{Method \code{update_hessian()}}{
Numerical computation of the Hessian matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GARCH_modelR6$update_hessian(coefficients, logLik, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coefficients}}{Numeric, named vector. Model's coefficients. When missing will be used the fitted parameters.}

\item{\code{logLik}}{Function, log-likelihood function depending on the parameters and x.}

\item{\code{...}}{Other arguments passed to \code{logLik} function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GARCH_modelR6-update_std.errors"></a>}}
\if{latex}{\out{\hypertarget{method-GARCH_modelR6-update_std.errors}{}}}
\subsection{Method \code{update_std.errors()}}{
Numerical computation of the std. errors of the parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GARCH_modelR6$update_std.errors(std.errors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{std.errors}}{Numeric std. errors.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GARCH_modelR6-next_step"></a>}}
\if{latex}{\out{\hypertarget{method-GARCH_modelR6-next_step}{}}}
\subsection{Method \code{next_step()}}{
Next step GARCH std. deviation forecast
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GARCH_modelR6$next_step(x = 1, sigma = 1, n.ahead = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Numeric, vector. Past residuals.}

\item{\code{sigma}}{Numeric, vector. Past garch std. deviations.}

\item{\code{n.ahead}}{Numeric, scalar. Number of steps ahead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GARCH_modelR6-print"></a>}}
\if{latex}{\out{\hypertarget{method-GARCH_modelR6-print}{}}}
\subsection{Method \code{print()}}{
Print method for \code{GARCH_modelR6} class.
Manual fit of the GARCH model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GARCH_modelR6$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GARCH_modelR6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GARCH_modelR6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GARCH_modelR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
