% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianMixture-R6.R
\name{gaussianMixture}
\alias{gaussianMixture}
\title{Gaussian mixture}
\description{
Gaussian mixture

Gaussian mixture
}
\details{
Fit the parameters of a gaussian mixture with k-components.

Applied after updating the parameters

Applied after updating the parameters
}
\note{
Version 1.0.0
}
\examples{
means = c(0.5,2)
sd = rep(1, 2)
p = c(0.2,  0.8)
# Grid
grid <- seq(-4, 4, 0.01)
plot(dmixnorm(grid, means, sd, p))
# Simulated sample
x <- rmixnorm(5000, means, sd, p)
# Gaussian mixture model
gm <- gaussianMixture$new(components=2)
# Fit the model
gm$fit(x$X)
gm
self <- gm$.__enclos_env__$self
private <- gm$.__enclos_env__$private
# EM-algo
gm$EM(x$X)
# Model parameters
gm$coefficients
# Fitted series
gm$fitted
# Theoric moments
gm$moments
gm$update(means = c(-2, 0, 2))
}
\keyword{gaussianMixture}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{maxit}}{Integer, maximum number of iterations.}

\item{\code{maxrestarts}}{Integer, maximum number of restarts.}

\item{\code{abstol}}{Numeric, absolute level for convergence.}

\item{\code{components}}{Integer, number of mixture components.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{means}}{Numeric vector containing the location parameter for each component.}

\item{\code{sd}}{Numeric vector containing the scale parameter for each component.}

\item{\code{p}}{Numeric vector containing the probability for each component.}

\item{\code{coefficients}}{named list with mixture coefficients.}

\item{\code{use_empiric}}{logical to denote if empiric parameters are currently used}

\item{\code{std.errors}}{named list with mixture parameters.}

\item{\code{model}}{Tibble with mixture parameters, in order means, sd, p.}

\item{\code{loglik}}{log-likelihood of the fitted series.}

\item{\code{fitted}}{fitted series}

\item{\code{moments}}{Tibble with the theoric moments and the number of observations used for fit.}

\item{\code{summary}}{Tibble with estimated parameters, std.errors and statistics}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gaussianMixture-new}{\code{gaussianMixture$new()}}
\item \href{#method-gaussianMixture-logLik}{\code{gaussianMixture$logLik()}}
\item \href{#method-gaussianMixture-E_step}{\code{gaussianMixture$E_step()}}
\item \href{#method-gaussianMixture-classify}{\code{gaussianMixture$classify()}}
\item \href{#method-gaussianMixture-fit}{\code{gaussianMixture$fit()}}
\item \href{#method-gaussianMixture-EM}{\code{gaussianMixture$EM()}}
\item \href{#method-gaussianMixture-update}{\code{gaussianMixture$update()}}
\item \href{#method-gaussianMixture-update_logLik}{\code{gaussianMixture$update_logLik()}}
\item \href{#method-gaussianMixture-update_empiric_parameters}{\code{gaussianMixture$update_empiric_parameters()}}
\item \href{#method-gaussianMixture-filter}{\code{gaussianMixture$filter()}}
\item \href{#method-gaussianMixture-Hessian}{\code{gaussianMixture$Hessian()}}
\item \href{#method-gaussianMixture-use_empiric_parameters}{\code{gaussianMixture$use_empiric_parameters()}}
\item \href{#method-gaussianMixture-print}{\code{gaussianMixture$print()}}
\item \href{#method-gaussianMixture-clone}{\code{gaussianMixture$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussianMixture-new"></a>}}
\if{latex}{\out{\hypertarget{method-gaussianMixture-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{gaussianMixture} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussianMixture$new(
  components = 2,
  maxit = 5000,
  maxrestarts = 500,
  abstol = 1e-08
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{components}}{(\code{integer(1)}), number of components.}

\item{\code{maxit}}{(\code{integer(1)}) Numeric, maximum number of iterations.}

\item{\code{maxrestarts}}{(\code{integer(1)}) Numeric, maximum number of restarts.}

\item{\code{abstol}}{(\code{numeric(1)}) Numeric, absolute level for convergence.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussianMixture-logLik"></a>}}
\if{latex}{\out{\hypertarget{method-gaussianMixture-logLik}{}}}
\subsection{Method \code{logLik()}}{
Compute the log-likelihood.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussianMixture$logLik(x, params, per_obs = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector}

\item{\code{params}}{Optional. Named list with mixture parameters.}

\item{\code{per_obs}}{Logical, when \code{TRUE} the log-likelihood is returned per observation,
otherwise is summed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussianMixture-E_step"></a>}}
\if{latex}{\out{\hypertarget{method-gaussianMixture-E_step}{}}}
\subsection{Method \code{E_step()}}{
Compute the posterior probabilities (E-step),
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussianMixture$E_step(x, params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Time series to fit.}

\item{\code{params}}{A named list with mixture parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussianMixture-classify"></a>}}
\if{latex}{\out{\hypertarget{method-gaussianMixture-classify}{}}}
\subsection{Method \code{classify()}}{
Classify the time series in the components with highest likelihood.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussianMixture$classify(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Time series to fit.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussianMixture-fit"></a>}}
\if{latex}{\out{\hypertarget{method-gaussianMixture-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the parameters with mclust package
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussianMixture$fit(
  x,
  weights,
  method = "mixtools",
  mu_target = NA,
  var_target = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector}

\item{\code{weights}}{observations weights, if a weight is equal to zero the observation is excluded, otherwise is included with unitary weight.}

\item{\code{method}}{Character, package used to fit the parameters. Can be \code{mclust} or \code{mixtools}.}

\item{\code{mu_target}}{Numeric, target mean of the mixture to match.}

\item{\code{var_target}}{Numeric, target variance of the mixture to match.
When \code{missing} all the available observations will be used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussianMixture-EM"></a>}}
\if{latex}{\out{\hypertarget{method-gaussianMixture-EM}{}}}
\subsection{Method \code{EM()}}{
Fit the parameters (EM-algorithm)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussianMixture$EM(x, weights)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector}

\item{\code{weights}}{observations weights, if a weight is equal to zero the observation is excluded, otherwise is included with unitary weight.
When \code{missing} all the available observations will be used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussianMixture-update"></a>}}
\if{latex}{\out{\hypertarget{method-gaussianMixture-update}{}}}
\subsection{Method \code{update()}}{
Update the parameters inside the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussianMixture$update(means, sd, p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{means}}{Numeric vector, means parameters.}

\item{\code{sd}}{Numeric vector, std. deviation parameters.}

\item{\code{p}}{Numeric vector, probabilities.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussianMixture-update_logLik"></a>}}
\if{latex}{\out{\hypertarget{method-gaussianMixture-update_logLik}{}}}
\subsection{Method \code{update_logLik()}}{
Update the log-likelihood with the current parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussianMixture$update_logLik()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussianMixture-update_empiric_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-gaussianMixture-update_empiric_parameters}{}}}
\subsection{Method \code{update_empiric_parameters()}}{
Compute the parameters on the classified time series.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussianMixture$update_empiric_parameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussianMixture-filter"></a>}}
\if{latex}{\out{\hypertarget{method-gaussianMixture-filter}{}}}
\subsection{Method \code{filter()}}{
Update the responsibilities, the log-likelihood, classify again the points and recompute empiric parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussianMixture$filter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussianMixture-Hessian"></a>}}
\if{latex}{\out{\hypertarget{method-gaussianMixture-Hessian}{}}}
\subsection{Method \code{Hessian()}}{
Hessian matrix \code{gaussianMixture} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussianMixture$Hessian()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussianMixture-use_empiric_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-gaussianMixture-use_empiric_parameters}{}}}
\subsection{Method \code{use_empiric_parameters()}}{
Substitute the empiric parameters with EM parameters. If evaluated again
the EM parameters will be substituted back.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussianMixture$use_empiric_parameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussianMixture-print"></a>}}
\if{latex}{\out{\hypertarget{method-gaussianMixture-print}{}}}
\subsection{Method \code{print()}}{
Print method for \code{gaussianMixture} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussianMixture$print(label)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{Character, optional label.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussianMixture-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gaussianMixture-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussianMixture$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
