% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGrid.R
\name{havDistance}
\alias{havDistance}
\title{Haversine distance}
\usage{
havDistance(lat_1, lon_1, lat_2, lon_2)
}
\arguments{
\item{lat_1}{Numeric vector, latitudes of first location.}

\item{lon_1}{Numeric vector, longitudes of first location.}

\item{lat_2}{Numeric vector, latitudes of second location.}

\item{lon_2}{Numeric vector, longitudes of second location.}
}
\value{
Vector of distances in kilometers.
}
\description{
Compute the Haversine distance between two points.
}
\examples{
havDistance(43.3, 12.1, 43.4, 12.2)
havDistance(c(43.35, 43.35), c(12.15, 12.1), c(43.4, 44.5), c(12.2, 13.4))
}
