% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGrid.R
\name{IDW}
\alias{IDW}
\title{Inverse Distance Weighting Functions}
\usage{
IDW(beta, d0)
}
\arguments{
\item{beta}{parameter used in exponential and power functions.}

\item{d0}{parameter used only in exponential function.}
}
\description{
Return a distance weighting function
}
\details{
When the parameter \code{d0} is not specified the function returned will be of power type otherwise of exponential type.
}
\examples{
# Power weighting
IDW_pow <- IDW(2)
IDW_pow(c(2, 3,10))
IDW_pow(c(2, 3,10), normalize = TRUE)
# Exponential weighting
IDW_exp <- IDW(2, d0 = 5)
IDW_exp(c(2, 3,10))
IDW_exp(c(2, 3,10), normalize = TRUE)
}
