% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelRegression-R6.R
\name{kernelRegression}
\alias{kernelRegression}
\title{Kernel regression}
\description{
Kernel regression

Kernel regression
}
\details{
Fit a kernel regression.
}
\note{
Version 1.0.0
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{model}}{an object of the class \code{npreg}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-kernelRegression-fit}{\code{kernelRegression$fit()}}
\item \href{#method-kernelRegression-predict}{\code{kernelRegression$predict()}}
\item \href{#method-kernelRegression-clone}{\code{kernelRegression$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kernelRegression-fit"></a>}}
\if{latex}{\out{\hypertarget{method-kernelRegression-fit}{}}}
\subsection{Method \code{fit()}}{
Fit a \code{kernelRegression} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kernelRegression$fit(formula, data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{formula, an object of class \code{formula} (or one that can be coerced to that class).}

\item{\code{data}}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.
If not found in data, the variables are taken from environment(formula), typically the environment from which \code{lm} is called.}

\item{\code{...}}{other parameters to be passed to the function \code{\link[np:np.regression]{np::npreg()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kernelRegression-predict"></a>}}
\if{latex}{\out{\hypertarget{method-kernelRegression-predict}{}}}
\subsection{Method \code{predict()}}{
Predict method for \code{kernelRegression} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kernelRegression$predict(newdata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kernelRegression-clone"></a>}}
\if{latex}{\out{\hypertarget{method-kernelRegression-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kernelRegression$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
