% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_test.R
\name{ks_test_ts}
\alias{ks_test_ts}
\title{Two sample Kolmogorov Smirnov test for a time series}
\usage{
ks_test_ts(
  x,
  ci = 0.05,
  idx_split,
  min_quantile = 0.015,
  max_quantile = 0.985,
  seed = 1,
  plot = FALSE
)
}
\arguments{
\item{x}{a vector.}

\item{ci}{p.value for rejection.}

\item{idx_split}{Index used for splitting the time series. If \code{missing} will be random sampled.}

\item{min_quantile}{minimum quantile for the grid of values.}

\item{max_quantile}{maximum quantile for the grid of values.}

\item{seed}{random seed for two sample test.}

\item{plot}{when \code{TRUE} a plot is returned, otherwise a \code{tibble}.}
}
\description{
Perform a two sample invariance test for a time series.
}
