% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik_std.errors}
\alias{logLik_std.errors}
\title{Compute the standard errors and robust standard errors}
\usage{
logLik_std.errors(params, logLik, robust = TRUE)
}
\description{
Compute the standard errors and robust standard errors
}
