% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riccati_root.R
\name{makeSemiPositive}
\alias{makeSemiPositive}
\title{Make a matrix positive semi-definite}
\usage{
makeSemiPositive(x, neg_values = 1e-05)
}
\arguments{
\item{x}{matrix, squared and symmetric.}

\item{neg_values}{numeric, the eigenvalues lower the zero will be substituted with this value.}
}
\description{
Make a matrix positive semi-definite
}
\examples{
m <- matrix(c(2, 2.99, 1.99, 2), nrow = 2, byrow = TRUE)
makeSemiPositive(m)

}
