% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyParams-R6.R
\name{monthlyParams}
\alias{monthlyParams}
\title{Create a function of time for monthly parameters}
\description{
Create a function of time for monthly parameters

Create a function of time for monthly parameters
}
\note{
Version 1.0.0
}
\examples{
set.seed(1)
params <- runif(12)
mp <- monthlyParams$new(params)
t_now <- as.Date("2022-01-01")
t_hor <- as.Date("2024-12-31")
dates <- seq.Date(t_now, t_hor, by = "1 day")
plot(mp$predict(dates), type = "l")
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{parameters}}{vector of parameters with length 12.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-monthlyParams-new}{\code{monthlyParams$new()}}
\item \href{#method-monthlyParams-predict}{\code{monthlyParams$predict()}}
\item \href{#method-monthlyParams-update}{\code{monthlyParams$update()}}
\item \href{#method-monthlyParams-clone}{\code{monthlyParams$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-monthlyParams-new"></a>}}
\if{latex}{\out{\hypertarget{method-monthlyParams-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{monthlyParams} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{monthlyParams$new(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{numeric vector of parameters with length 12.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-monthlyParams-predict"></a>}}
\if{latex}{\out{\hypertarget{method-monthlyParams-predict}{}}}
\subsection{Method \code{predict()}}{
Predict the monthly paramete
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{monthlyParams$predict(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{date as character or month as numeric.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-monthlyParams-update"></a>}}
\if{latex}{\out{\hypertarget{method-monthlyParams-update}{}}}
\subsection{Method \code{update()}}{
Update the monthly parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{monthlyParams$update(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{numeric vector of parameters with length 12.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-monthlyParams-clone"></a>}}
\if{latex}{\out{\hypertarget{method-monthlyParams-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{monthlyParams$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
