% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvgaussianMixture.R
\name{mvgaussianMixture}
\alias{mvgaussianMixture}
\title{Multivariate gaussian mixture}
\usage{
mvgaussianMixture(
  x,
  means,
  sd,
  p,
  components = 2,
  maxit = 100,
  abstol = 1e-14,
  na.rm = FALSE
)
}
\description{
Multivariate gaussian mixture
}
