% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{number_of_day}
\alias{number_of_day}
\title{Number of day}
\usage{
number_of_day(x)
}
\arguments{
\item{x}{dates vector in the format \code{YYYY-MM-DD}.}
}
\value{
Numeric vector with the number of the day during the year.
}
\description{
Compute the number of day of the year given a vector of dates.
}
\examples{
number_of_day("2040-01-31")
number_of_day(c("2015-03-31", "2016-03-31", "2017-03-31"))
number_of_day(c("2015-02-28", "2016-02-28", "2017-02-28"))
number_of_day(c("2015-03-01", "2016-03-01", "2017-03-01"))
}
