% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riccati_root.R
\name{riccati_root}
\alias{riccati_root}
\title{Riccati Root}
\usage{
riccati_root(x)
}
\arguments{
\item{x}{squared and symmetric matrix.}
}
\description{
Compute the square root of a symmetric matrix.
}
\examples{
cv <- matrix(c(1, 0.3, 0.3, 1), nrow = 2, byrow = TRUE)
riccati_root(cv)

}
