% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalSolarFunctions-R6.R
\name{seasonalSolarFunctions}
\alias{seasonalSolarFunctions}
\title{Solar seasonal functions}
\description{
Solar seasonal functions

Solar seasonal functions
}
\note{
Version 1.0.0
}
\examples{
dates <- seq.Date(as.Date("2022-01-01"), as.Date("2022-12-31"), 1)
# Seasonal functions object
sf <- seasonalSolarFunctions$new()

# Adjustment parameter
sf$B(number_of_day(dates))

# Time adjustment in minutes
sf$E(dates)

# Declination
sf$declination(dates)

# Solar constant
sf$Gsc

# Solar constant adjusted
sf$Gon(dates)

# Extraterrestrial radiation
sf$Hon(dates, 43)

# Number of hours of sun
sf$sun_hours(dates, 43)

# Sunset hour angle
sf$sunset_hour_angle(dates, 43)

sf$solar_time("2022-01-01 12:00", 11, 10)
sf$hour_angle("2022-01-01 14:00", 11, 15)
sf$incidence_angle("2022-06-01 21:00", 31, 12, lon_st = 15, beta = 0, gamma = 0)
sf$azimut_angle("2022-01-01 14:00", 30, 17, lon_st = 15)

}
\references{
Duffie, Solar Engineering of Thermal Processes Fourth Edition.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{legal_hour}}{Logical, when \code{TRUE} the clock time will be corrected for the legal hour.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{Gsc}}{solar constant, i,e, \code{1367}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-seasonalSolarFunctions-new}{\code{seasonalSolarFunctions$new()}}
\item \href{#method-seasonalSolarFunctions-update_method}{\code{seasonalSolarFunctions$update_method()}}
\item \href{#method-seasonalSolarFunctions-B}{\code{seasonalSolarFunctions$B()}}
\item \href{#method-seasonalSolarFunctions-degree}{\code{seasonalSolarFunctions$degree()}}
\item \href{#method-seasonalSolarFunctions-radiant}{\code{seasonalSolarFunctions$radiant()}}
\item \href{#method-seasonalSolarFunctions-E}{\code{seasonalSolarFunctions$E()}}
\item \href{#method-seasonalSolarFunctions-elevation}{\code{seasonalSolarFunctions$elevation()}}
\item \href{#method-seasonalSolarFunctions-solar_time}{\code{seasonalSolarFunctions$solar_time()}}
\item \href{#method-seasonalSolarFunctions-solar_hour}{\code{seasonalSolarFunctions$solar_hour()}}
\item \href{#method-seasonalSolarFunctions-hour_angle}{\code{seasonalSolarFunctions$hour_angle()}}
\item \href{#method-seasonalSolarFunctions-incidence_angle}{\code{seasonalSolarFunctions$incidence_angle()}}
\item \href{#method-seasonalSolarFunctions-azimut_angle}{\code{seasonalSolarFunctions$azimut_angle()}}
\item \href{#method-seasonalSolarFunctions-Gon}{\code{seasonalSolarFunctions$Gon()}}
\item \href{#method-seasonalSolarFunctions-declination}{\code{seasonalSolarFunctions$declination()}}
\item \href{#method-seasonalSolarFunctions-Hon}{\code{seasonalSolarFunctions$Hon()}}
\item \href{#method-seasonalSolarFunctions-sunset_hour_angle}{\code{seasonalSolarFunctions$sunset_hour_angle()}}
\item \href{#method-seasonalSolarFunctions-sun_hours}{\code{seasonalSolarFunctions$sun_hours()}}
\item \href{#method-seasonalSolarFunctions-solar_altitude}{\code{seasonalSolarFunctions$solar_altitude()}}
\item \href{#method-seasonalSolarFunctions-solar_angles}{\code{seasonalSolarFunctions$solar_angles()}}
\item \href{#method-seasonalSolarFunctions-clearsky}{\code{seasonalSolarFunctions$clearsky()}}
\item \href{#method-seasonalSolarFunctions-clone}{\code{seasonalSolarFunctions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-new"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{seasonalSolarFunctions} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$new(method = "spencer", legal_hour = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{character, method type for computations. Can be \code{cooper} or \code{spencer}.}

\item{\code{legal_hour}}{Logical, when \code{TRUE} the clock time will be corrected for the legal hour.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-update_method"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-update_method}{}}}
\subsection{Method \code{update_method()}}{
Extract or update the method used for computations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$update_method(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{character, method type. Can be \code{cooper} or \code{spencer}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
When \code{x} is missing it return a character containing the method that
is actually used.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-B"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-B}{}}}
\subsection{Method \code{B()}}{
Seasonal adjustment parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$B(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function implement Eq. 1.4.2 from Duffie (4th edition), i.e.
\deqn{B(n) = \frac{2\pi}{365} n}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-degree"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-degree}{}}}
\subsection{Method \code{degree()}}{
Convert angles in radiant into an angles in degrees.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$degree(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric vector, angles in radiant.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{\frac{x 180}{\pi}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-radiant"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-radiant}{}}}
\subsection{Method \code{radiant()}}{
Convert angles in degrees into an angles in radiant
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$radiant(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric vector, angles in degrees.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{\frac{x \pi}{180}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-E"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-E}{}}}
\subsection{Method \code{E()}}{
Compute the time adjustment in minutes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$E(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function implement Eq. 1.5.3 from Duffie (4th edition), i.e.
\deqn{E = 229.2(0.000075 + 0.001868 \cos(B) - 0.032077\sin(B) - 0.014615\cos(2B) - 0.04089\sin(2B))}
}

\subsection{Returns}{
The time adjustment in minutes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-elevation"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-elevation}{}}}
\subsection{Method \code{elevation()}}{
Compute the angle in the degree given a certain altitude in meters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$elevation(alt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alt}}{Numeric, altitude in meters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-solar_time"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-solar_time}{}}}
\subsection{Method \code{solar_time()}}{
Compute the solar time from a clock time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$solar_time(x, lon, lon_st = 15, tz = "Europe/Rome")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{datetime, clock hour.}

\item{\code{lon}}{longitude of interest in degrees.}

\item{\code{lon_st}}{longitude of the Local standard meridian in degrees.}

\item{\code{tz}}{Character, reference time zone.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function implement Eq. 1.5.2 from Duffie (4th edition), i.e.
\deqn{solartime = clocktime + 4 (lon-lon_{st}) + E(n)}
}

\subsection{Returns}{
A datetime object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-solar_hour"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-solar_hour}{}}}
\subsection{Method \code{solar_hour()}}{
Compute the solar hour for a specific clock time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$solar_hour(LST)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{LST}}{datetime, true solar time.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Hours
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-hour_angle"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-hour_angle}{}}}
\subsection{Method \code{hour_angle()}}{
Compute the solar angle for a specific hour of the day.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$hour_angle(LST)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{LST}}{datetime, true solar time.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function implement Eq. 1.42 from Comini (2013), i.e.
\deqn{\omega = 15 (solarhour - 12)}
where the "solarhour" is expressed in hours.
}

\subsection{Returns}{
An angle in degrees
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-incidence_angle"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-incidence_angle}{}}}
\subsection{Method \code{incidence_angle()}}{
Compute the incidence angle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$incidence_angle(LST, lat, alt = 0, beta = 0, gamma = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{LST}}{datetime, true solar time.}

\item{\code{lat}}{latitude of interest in degrees.}

\item{\code{alt}}{Numeric, altitude in meters.}

\item{\code{beta}}{altitude}

\item{\code{gamma}}{orientation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An angle in degrees
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-azimut_angle"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-azimut_angle}{}}}
\subsection{Method \code{azimut_angle()}}{
Compute the solar azimuth angle for a specific time of the day.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$azimut_angle(LST, lat, alt, beta = 0, gamma = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{LST}}{datetime, true solar time.}

\item{\code{lat}}{latitude of interest in degrees.}

\item{\code{alt}}{Numeric, altitude in meters.}

\item{\code{beta}}{altitude}

\item{\code{gamma}}{orientation}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function implement Eq. 1.6.6 from Duffie (4th edition), i.e.
\deqn{\gamma_s = sign(\omega) \left|\cos^{-1}\left( \frac{\cos \theta_z \sin \phi - \sin \delta}{\sin \theta_z \cos \phi} \right) \right|}
}

\subsection{Returns}{
The solar azimut angle in degrees
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-Gon"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-Gon}{}}}
\subsection{Method \code{Gon()}}{
Compute the solar constant adjusted for the day of the year.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$Gon(n, deriv = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year.}

\item{\code{deriv}}{Logical, when \code{TRUE} will return the first derivative with respect to time.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
When method is \code{cooper} the function implement Eq. 1.4.1a from Duffie (4th edition), i.e.
\deqn{G_{o,n} = G_{sc} (1 + 0.033\cos(B))}
otherwise when it is \code{spencer} it implement Eq. 1.4.1b from Duffie (4th edition):
\deqn{G_{o,n} = G_{sc} (1.000110 + 0.034221\cos(B) + 0.001280\sin(B) + 0.000719\cos(2B) + 0.000077\sin(2B))}
When \code{deriv = TRUE} it will be returned the derivatives with respect to time. When the method is \code{cooper}:
\deqn{\frac{\partial G_{o,n}}{\partial n} = - G_{sc} \frac{2\pi}{365} 0.033 \sin(B))}
Otherwise if it is \code{spencer}:
\deqn{\frac{\partial G_{o,n}}{\partial n} = G_{sc} \frac{2\pi}{365} (-0.034221\sin(B) + 0.001280\cos(B) - 0.001438\sin(2B) + 0.000154\cos(2B))}
}

\subsection{Returns}{
The solar constant in \eqn{W/m^2} for the day n.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-declination"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-declination}{}}}
\subsection{Method \code{declination()}}{
Compute solar declination in degrees.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$declination(n, deriv = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}

\item{\code{deriv}}{Logical, when \code{TRUE} will return the first derivative with respect to time.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
When method is \code{cooper} the function implement Eq. 1.6.1a from Duffie (4th edition), i.e.
\deqn{\delta(n) = 23.45 \sin \left(\frac{2 \pi (284 + n)}{365}\right)}
otherwise when it is \code{spencer} it implement Eq. 1.6.1b from Duffie (4th edition):
\deqn{\delta(n) = \frac{180}{\pi}(0.006918 - 0.399912\cos(B) + 0.070257\sin(B) - 0.006758\cos(2B) + 0.000907\sin(2B) - 0.002697\cos(3B) + 0.00148\sin(3B))}
When \code{deriv = TRUE} it will be returned the derivatives with respect to time. When the method is \code{cooper}:
\deqn{\frac{\partial \delta}{\partial n}(n) = 23.45 \frac{2\pi}{365} \cos \left(\frac{2 \pi (284 + n)}{365}\right)}
otherwise when the method is \code{spencer}:
\deqn{\frac{\partial \delta}{\partial n}(n) = \frac{360}{365}(0.399912\sin(B) + 0.070257\cos(B) + 0.013516\sin(2B) + 0.001814\cos(2B) + 0.008091\sin(3B) + 0.00444\cos(3B))}
}

\subsection{Returns}{
The solar declination in degrees.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-Hon"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-Hon}{}}}
\subsection{Method \code{Hon()}}{
Compute the solar extraterrestrial radiation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$Hon(n, lat, alt, deriv = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}

\item{\code{lat}}{latitude of interest in degrees.}

\item{\code{alt}}{Numeric, altitude in meters.}

\item{\code{deriv}}{Logical, when \code{TRUE} will return the first derivative with respect to time.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function implement Eq. 1.10.3 from Duffie (4th edition):
\deqn{H_{on} = G_{on} \frac{24 \times 3600}{\pi} (\cos(lat) \cos(\delta) \sin(\omega_s) + \frac{\pi}{180}\sin(lat) \sin(\delta))}
}

\subsection{Returns}{
Extraterrestrial radiation on an horizontal surface in kilowatt hour for meters squared for day.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-sunset_hour_angle"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-sunset_hour_angle}{}}}
\subsection{Method \code{sunset_hour_angle()}}{
Compute solar angle at sunset in degrees
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$sunset_hour_angle(n, lat, alt, deriv = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}

\item{\code{lat}}{Numeric, latitude of interest in degrees.}

\item{\code{alt}}{Numeric, altitude in meters.}

\item{\code{deriv}}{Logical, when \code{TRUE} will return the first derivative with respect to time.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function implement Eq. 1.6.10 from Duffie (4th edition), i.e.
\deqn{\omega_s = \cos^{-1}(-\tan(\delta(n))\tan(\phi))}
When altitude is not missing it will implement a generalized version with altitude, i.e.
\deqn{\omega_s = \cos^{-1}\left(\frac{\sin H - \sin\delta \sin\phi}{\cos\phi \cos \delta}\right)}
}

\subsection{Returns}{
The sunset hour angle in degrees.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-sun_hours"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-sun_hours}{}}}
\subsection{Method \code{sun_hours()}}{
Compute number of sun hours for a day n.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$sun_hours(n, lat, alt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year.}

\item{\code{lat}}{Numeric, latitude of interest in degrees.}

\item{\code{alt}}{Numeric, altitude in meters.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function implement Eq. 1.6.11 from Duffie (4th edition), i.e.
\deqn{\frac{2}{15} \omega_s}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-solar_altitude"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-solar_altitude}{}}}
\subsection{Method \code{solar_altitude()}}{
Compute solar altitude in degrees
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$solar_altitude(n, lat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}

\item{\code{lat}}{Numeric, latitude of interest in degrees.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes
\deqn{\sin^{-1}(-\sin(\delta(n))\sin(\phi) + \cos(\delta(n))\cos(\phi))}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-solar_angles"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-solar_angles}{}}}
\subsection{Method \code{solar_angles()}}{
Compute the solar angle for a latitude in different dates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$solar_angles(
  x,
  lat,
  lon,
  alt,
  lon_st = 15,
  beta = 0,
  gamma = 0,
  by = "1 min",
  tz = "Europe/Rome"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{datetime, clock hour.}

\item{\code{lat}}{Numeric, latitude of interest in degrees.}

\item{\code{lon}}{Numeric, longitude of interest in degrees.}

\item{\code{alt}}{Numeric, altitude in meters.}

\item{\code{lon_st}}{Numeric, longitude of the Local standard meridian in degrees}

\item{\code{beta}}{Numeric angle, inclination of the solar panel.}

\item{\code{gamma}}{Numeric, angle orientation of the panel.}

\item{\code{by}}{Character, time step. Default is \verb{1 min}.}

\item{\code{tz}}{Character, reference time zone.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-clearsky"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-clearsky}{}}}
\subsection{Method \code{clearsky()}}{
Hottel clearsky
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$clearsky(
  cosZ = NULL,
  G0 = NULL,
  alt,
  clime = "No Correction"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cosZ}}{solar incidence angle}

\item{\code{G0}}{solar constant}

\item{\code{alt}}{Numeric, altitude in meters.}

\item{\code{clime}}{clime correction}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalSolarFunctions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalSolarFunctions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
