% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sGARCH.R
\name{sGARCH}
\alias{sGARCH}
\title{Implementation of \code{rugarch} methods for a GARCH(p,q) as R6 class}
\description{
Implementation of \code{rugarch} methods for a GARCH(p,q) as R6 class

Implementation of \code{rugarch} methods for a GARCH(p,q) as R6 class
}
\note{
Version 1.0.0
}
\keyword{GARCH}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{archOrder}}{Numeric scalar, ARCH order.}

\item{\code{garchOrder}}{Numeric scalar, GARCH order.}

\item{\code{order}}{Numeric named vector, orders of the GARCH model. The first element is the ARCH order, while the second the GARCH order.}

\item{\code{omega}}{Numeric scalar, intercept parameter.}

\item{\code{alpha}}{Numeric vector, ARCH parameters.}

\item{\code{beta}}{Numeric vector, GARCH parameters.}

\item{\code{coefficients}}{Numeric vector, model's coefficients.}

\item{\code{A}}{Numeric matrix, companion matrix.}

\item{\code{b}}{Numeric vector.}

\item{\code{d}}{Numeric vector}

\item{\code{std.errors}}{Numeric vector, standard errors of the model's coefficients.}

\item{\code{sigma2_inf}}{Numeric scalar, long-term unconditional std. deviation.}

\item{\code{loglik}}{model log-likelihood}

\item{\code{tidy}}{Tibble with estimated parameters and relative std. errors.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sGARCH-new}{\code{sGARCH$new()}}
\item \href{#method-sGARCH-fit}{\code{sGARCH$fit()}}
\item \href{#method-sGARCH-filter}{\code{sGARCH$filter()}}
\item \href{#method-sGARCH-update}{\code{sGARCH$update()}}
\item \href{#method-sGARCH-update_std.errors}{\code{sGARCH$update_std.errors()}}
\item \href{#method-sGARCH-next_step}{\code{sGARCH$next_step()}}
\item \href{#method-sGARCH-print}{\code{sGARCH$print()}}
\item \href{#method-sGARCH-clone}{\code{sGARCH$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sGARCH-new"></a>}}
\if{latex}{\out{\hypertarget{method-sGARCH-new}{}}}
\subsection{Method \code{new()}}{
Initialize a standard GARCH model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sGARCH$new(archOrder = 1, garchOrder = 1, mode = "unitOmega")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{archOrder}}{Integer scalar, ARCH order.}

\item{\code{garchOrder}}{Integer scalar, GARCH order.}

\item{\code{mode}}{Character, one of \code{"unitOmega"}, \code{"targetSigma2"}, \code{"freeOmega"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sGARCH-fit"></a>}}
\if{latex}{\out{\hypertarget{method-sGARCH-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the GARCH model with \code{rugarch} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sGARCH$fit(x, weights)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Numeric, vector. Time series to be fitted.}

\item{\code{weights}}{Numeric, vector. Optional custom weights.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sGARCH-filter"></a>}}
\if{latex}{\out{\hypertarget{method-sGARCH-filter}{}}}
\subsection{Method \code{filter()}}{
Filter method from \code{rugarch} package to compute GARCH variance, residuals and log-likelihoods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sGARCH$filter(x, eps0 = NULL, sigma20 = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Numeric, vector. Time series to be filtered.}

\item{\code{eps0}}{Optional numeric initial epsilons to prepend (length p+q).}

\item{\code{sigma20}}{Optional numeric initial variances to prepend (length p+q).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sGARCH-update"></a>}}
\if{latex}{\out{\hypertarget{method-sGARCH-update}{}}}
\subsection{Method \code{update()}}{
Update the coefficients of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sGARCH$update(coefficients)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coefficients}}{Numeric, named vector. Model's coefficients.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sGARCH-update_std.errors"></a>}}
\if{latex}{\out{\hypertarget{method-sGARCH-update_std.errors}{}}}
\subsection{Method \code{update_std.errors()}}{
Numerical computation of the std. errors of the parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sGARCH$update_std.errors(std.errors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{std.errors}}{Numeric std. errors.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sGARCH-next_step"></a>}}
\if{latex}{\out{\hypertarget{method-sGARCH-next_step}{}}}
\subsection{Method \code{next_step()}}{
Next step GARCH std. deviation forecast
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sGARCH$next_step(eps0, sigma20)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{eps0}}{Numeric initial epsilons to prepend (length p+q).}

\item{\code{sigma20}}{Numeric initial variances to prepend (length p+q).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sGARCH-print"></a>}}
\if{latex}{\out{\hypertarget{method-sGARCH-print}{}}}
\subsection{Method \code{print()}}{
Print method for \code{GARCH_modelR6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sGARCH$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sGARCH-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sGARCH-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sGARCH$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
