% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sGARCH-internals.R
\name{sGARCH_filter}
\alias{sGARCH_filter}
\title{Fast sGARCH filter (C)}
\usage{
sGARCH_filter(x, omega, alpha, beta, eps0 = NULL, sigma20 = NULL)
}
\arguments{
\item{x}{numeric residuals/innovations.}

\item{omega}{Numeric scalar, strictly positive, intercept of the GARCH model.}

\item{alpha}{Numeric scalar, strictly positive, ARCH parameters.}

\item{beta}{numeric length q (>=0).}

\item{eps0}{optional numeric initial epsilons to prepend (length p+q).}

\item{sigma20}{optional numeric initial variances to prepend (length p+q).}
}
\value{
numeric vector of conditional variances sigma2, same length as the working series (prepended if initials are provided).
}
\description{
Fast sGARCH filter (C)
}
\examples{
# Simple GARCH(1,1)
set.seed(1)
x <- rnorm(1000)
omega <- 0.1; alpha <- 0.05; beta <- 0.9
sGARCH_filter(x, omega, alpha, beta)  # no initials

# With initials (k=max(p,q)=1)
sGARCH_filter(x, omega, alpha, beta, eps0 = 0, sigma20 = 0.2)

}
