% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sGARCH-internals.R
\name{sGARCH_params_to_zeta}
\alias{sGARCH_params_to_zeta}
\title{Map constrained GARCH coefficients back to unconstrained parameters}
\usage{
sGARCH_params_to_zeta(phi, p, q, mode = "unitOmega", eps = 0)
}
\arguments{
\item{phi}{Named numeric vector \code{c(omega, alpha_1..alpha_p, beta_1..beta_q)}.}

\item{p}{Integer, ARCH order.}

\item{q}{Integer, GARCH order.}

\item{mode}{Character, one of \code{"unitOmega"}, \code{"targetSigma2"}, \code{"freeOmega"}.}

\item{eps}{Numeric buffer in [0,1), enforces \eqn{\tau \le 1-\varepsilon}.}
}
\value{
Numeric vector \code{(eta0, kappa_1..kappa_{p+q}, [xi|psi])}.
}
\description{
Inverts \code{phi = (omega, alpha_1..alpha_p, beta_1..beta_q)} into
\code{zeta = (eta0, kappa_1..kappa_{p+q}, [xi|psi])}, fixing the softmax gauge
by setting the last \code{kappa} to 0.
}
\examples{
phi <- c(omega = 0.1, alpha1 = 0.05, beta1 = 0.9)
GARCH_params_to_zeta(phi, p = 1, q = 1)

}
