% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarHedging.R
\name{solarHedging_model}
\alias{solarHedging_model}
\title{Global Minimum Variance number of contracts}
\usage{
solarHedging_model(
  model,
  moments,
  P0_P,
  r_star,
  gamma = 0.01,
  put = TRUE,
  control_options = control_solarOption(),
  control_hedge = control_solarHedging()
)
}
\arguments{
\item{model}{An object with the class \code{solarModel}. See the function \code{\link{solarModel}} for more details.}

\item{moments}{Tibble containing the forecasted moments for different days ahead. See the function \code{\link{solarMoments}} for more details.}

\item{P0_P}{Optional numeric scalar, expected value of 1 solar derivative with unitary tick under \eqn{\mathbb{P}}.}

\item{gamma}{Numeric scalar, risk aversion parameter.}

\item{put}{Logical, when \code{TRUE}, the default, will be computed the price for a \code{put} contract, otherwise for a \code{call} contract.}

\item{control_options}{Named list with control parameters. See \code{\link{control_solarOption}} for more details.}

\item{control_hedge}{Named list, control parameters for hedging. See the function \code{\link{control_solarHedging}} for more details.}
}
\description{
Compute the optimal number of contracts, such that the variance of the cash flow of a solar power producer is minimum.
}
\note{
Version 1.0.0.
}
\keyword{solarHedging}
