% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarHedging.R
\name{solarHedging_scenarios}
\alias{solarHedging_scenarios}
\title{Compute the optimal number of solar derivative index}
\usage{
solarHedging_scenarios(
  scenarios,
  P0_P,
  r_star = 0,
  gamma = 1e-04,
  put = TRUE,
  control_options = control_solarOption(),
  control_hedge = control_solarHedging()
)
}
\arguments{
\item{scenarios}{Object with the class \code{solarScenario}. See the function \code{\link{solarScenario}} for more details.}

\item{P0_P}{Optional numeric scalar, expected value of 1 solar derivative with unitary tick under \eqn{\mathbb{P}}.}

\item{gamma}{Numeric scalar, risk aversion parameter.}

\item{put}{Logical, when \code{TRUE}, the default, will be computed the price for a \code{put} contract, otherwise for a \code{call} contract.}

\item{control_options}{Named list with control parameters. See \code{\link{control_solarOption}} for more details.}

\item{control_hedge}{Named list, control parameters for hedging. See the function \code{\link{control_solarHedging}} for more details.}
}
\description{
Compute the optimal number of contracts, such that the variance of the cash flow of a solar power
producer with a given setup is minimum.
}
\note{
Version 1.0.0.
}
\keyword{solarHedging}
