% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-VaR.R
\name{solarMixture_VaR}
\alias{solarMixture_VaR}
\title{Compute the Value at Risk and Expected Shortfall of a SolarMixture}
\usage{
solarMixture_VaR(solarMix, date, x, alpha = 0.05, ci = 0.05, ES = FALSE)
}
\arguments{
\item{alpha}{Numeric vector of confidence levels. Allows for more than one alpha.}

\item{ci}{Numeric scalar, confidence levels used to state if the Null is rejected or not on VaR tests.}

\item{ES}{Logical, when \code{TRUE} the expected shortfall will be also computed for each alpha.}

\item{model}{solarMixture}

\item{type}{Numeric, type of evaluation, \code{full} on the complete data, \code{train} on the train data, \code{test} on the test data.}
}
\description{
Compute the Value at Risk and Expected Shortfall of a SolarMixture
}
