% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-R6.R
\name{solarModel}
\alias{solarModel}
\title{Solar Model in R6 Class}
\description{
The \code{solarModel} class allows for the step-by-step fitting and transformation of solar radiation data, from clear sky
models to GARCH models for residual analysis. It utilizes various private and public methods to fit the seasonal
clearsky model, compute risk drivers, detect outliers, and apply time-series models.
}
\details{
The \code{solarModel} class is an implementation of a comprehensive solar model that includes fitting seasonal models,
detecting outliers, performing transformations, and applying time-series models such as AR and GARCH. This model
is specifically designed to predict solar radiation data, and it uses seasonal and Gaussian Mixture models to
capture the underlying data behavior.
}
\note{
Version 1.0.0.
}
\examples{

# Model specification
spec <- solarModel_spec$new()
spec$set_mean.model(arOrder = 1, maOrder = 1)
spec$specification("Bologna")
spec
# Model fit
Bologna <- solarModel$new(spec)
Bologna$fit()
Bologna
# save(spec, file = "data/Bologna.RData")

# Extract and update the parameters
model <- Bologna$clone(TRUE)
params <- model$coefficients
model$update(params)
model$filter()

# Fit a model with the realized clear sky
spec$control$stochastic_clearsky <- TRUE
# Initialize a new model
model <- solarModel$new(spec)
# Model fit
model$fit()

# Fit a model for the clearsky
spec_Ct <- spec
spec_Ct$control$stochastic_clearsky <- FALSE
spec_Ct$target <- "clearsky"
# Initialize a new model
model <- solarModel$new(spec)
# Model fit
model$fit()

}
\keyword{solarModel}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{place}}{Character, optional name of the location considered.}

\item{\code{model_name}}{Character, model's name.}

\item{\code{data}}{A data frame with the fitted data, and the seasonal and monthly parameters.}

\item{\code{seasonal_data}}{A data frame containing seasonal and monthly parameters.}

\item{\code{monthly_data}}{A data frame that contains monthly parameters.}

\item{\code{loglik}}{The log-likelihood computed on train data.}

\item{\code{spec}}{A list with the specification that govern the behavior of the model's fitting process.}

\item{\code{location}}{A data frame with coordinates of the location considered.}

\item{\code{transform}}{A \code{\link{solarTransform}} object with the transformation functions applied to the data.}

\item{\code{seasonal_model_Ct}}{The fitted model for clear sky radiation, used for predict the maximum radiation available.}

\item{\code{seasonal_model_Yt}}{The fitted seasonal model for the target variable.}

\item{\code{ARMA}}{The fitted ARMA model for the target variable.}

\item{\code{seasonal_variance}}{The fitted model for seasonal variance.}

\item{\code{GARCH}}{A model object representing the GARCH model fitted to the residuals.}

\item{\code{NM_model}}{A model object representing the Gaussian Mixture model fitted to the standardized residuals.}

\item{\code{moments}}{Get a list containing the conditional and unconditional moments.}

\item{\code{coefficients}}{Get the model parameters as a named list.}

\item{\code{var_theta}}{Variance-covariance matrix of the parameters with robust std. errors.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-solarModel-new}{\code{solarModel$new()}}
\item \href{#method-solarModel-fit}{\code{solarModel$fit()}}
\item \href{#method-solarModel-fit_seasonal_model_Ct}{\code{solarModel$fit_seasonal_model_Ct()}}
\item \href{#method-solarModel-compute_risk_drivers}{\code{solarModel$compute_risk_drivers()}}
\item \href{#method-solarModel-fit_transform}{\code{solarModel$fit_transform()}}
\item \href{#method-solarModel-fit_seasonal_model_Yt}{\code{solarModel$fit_seasonal_model_Yt()}}
\item \href{#method-solarModel-fit_monthly_mean}{\code{solarModel$fit_monthly_mean()}}
\item \href{#method-solarModel-fit_ARMA}{\code{solarModel$fit_ARMA()}}
\item \href{#method-solarModel-fit_seasonal_variance}{\code{solarModel$fit_seasonal_variance()}}
\item \href{#method-solarModel-fit_monthly_variance}{\code{solarModel$fit_monthly_variance()}}
\item \href{#method-solarModel-correct_seasonal_variance}{\code{solarModel$correct_seasonal_variance()}}
\item \href{#method-solarModel-fit_GARCH}{\code{solarModel$fit_GARCH()}}
\item \href{#method-solarModel-fit_NM_model}{\code{solarModel$fit_NM_model()}}
\item \href{#method-solarModel-update}{\code{solarModel$update()}}
\item \href{#method-solarModel-update_moments}{\code{solarModel$update_moments()}}
\item \href{#method-solarModel-update_logLik}{\code{solarModel$update_logLik()}}
\item \href{#method-solarModel-update_risk_drivers}{\code{solarModel$update_risk_drivers()}}
\item \href{#method-solarModel-update_NM_classification}{\code{solarModel$update_NM_classification()}}
\item \href{#method-solarModel-filter}{\code{solarModel$filter()}}
\item \href{#method-solarModel-Moments}{\code{solarModel$Moments()}}
\item \href{#method-solarModel-VaR}{\code{solarModel$VaR()}}
\item \href{#method-solarModel-logLik}{\code{solarModel$logLik()}}
\item \href{#method-solarModel-print}{\code{solarModel$print()}}
\item \href{#method-solarModel-clone}{\code{solarModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{solarModel}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$new(spec)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{spec}}{an object with class \code{solarModelSpec}. See the function \code{\link{solarModel_spec}} for details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-fit}{}}}
\subsection{Method \code{fit()}}{
Initialize and fit a \code{\link{solarModel}} object given the specification contained in \verb{$control}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-fit_seasonal_model_Ct"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-fit_seasonal_model_Ct}{}}}
\subsection{Method \code{fit_seasonal_model_Ct()}}{
Initialize and fit a \code{\link{seasonalClearsky}} model given the specification contained in \verb{$control}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$fit_seasonal_model_Ct()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-compute_risk_drivers"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-compute_risk_drivers}{}}}
\subsection{Method \code{compute_risk_drivers()}}{
Compute the risk drivers and impute the observation that are greater or equal to the clear sky level.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$compute_risk_drivers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-fit_transform"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-fit_transform}{}}}
\subsection{Method \code{fit_transform()}}{
Fit the parameters of the \code{\link{solarTransform}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$fit_transform()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-fit_seasonal_model_Yt"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-fit_seasonal_model_Yt}{}}}
\subsection{Method \code{fit_seasonal_model_Yt()}}{
Fit a \code{\link{seasonalModel}} the transformed variable (\code{Yt}) and compute deseasonalized series (\code{Yt_tilde}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$fit_seasonal_model_Yt()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-fit_monthly_mean"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-fit_monthly_mean}{}}}
\subsection{Method \code{fit_monthly_mean()}}{
Correct the deseasonalized series (\code{Yt_tilde}) by subtracting its monthly mean (\code{Yt_tilde_uncond}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$fit_monthly_mean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-fit_ARMA"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-fit_ARMA}{}}}
\subsection{Method \code{fit_ARMA()}}{
Fit an AR model (\code{Yt_tilde}) and compute AR residuals (\code{eps}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$fit_ARMA()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-fit_seasonal_variance"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-fit_seasonal_variance}{}}}
\subsection{Method \code{fit_seasonal_variance()}}{
Fit a \code{\link{seasonalModel}} on AR squared residuals (\code{eps}) and compute deseasonalized residuals \code{eps_tilde}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$fit_seasonal_variance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-fit_monthly_variance"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-fit_monthly_variance}{}}}
\subsection{Method \code{fit_monthly_variance()}}{
Correct the standardized series (\code{eps_tilde}) by subtracting its monthly mean (\code{sigma_uncond}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$fit_monthly_variance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-correct_seasonal_variance"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-correct_seasonal_variance}{}}}
\subsection{Method \code{correct_seasonal_variance()}}{
Correct the parameters of the seasonal variance to ensure a unitary variance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$correct_seasonal_variance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-fit_GARCH"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-fit_GARCH}{}}}
\subsection{Method \code{fit_GARCH()}}{
Fit a \code{GARCH} model on the deseasonalized residuals (\code{eps_tilde}).
Compute the standardized (\code{u}) and monthly deseasonalized residuals (\code{u_tilde}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$fit_GARCH()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-fit_NM_model"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-fit_NM_model}{}}}
\subsection{Method \code{fit_NM_model()}}{
Initialize and fit a \code{solarMixture} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$fit_NM_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-update"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-update}{}}}
\subsection{Method \code{update()}}{
Update the parameters inside object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$update(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{List of parameters. See the slot \verb{$coefficients} for a template.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-update_moments"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-update_moments}{}}}
\subsection{Method \code{update_moments()}}{
Update the moments inside object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$update_moments()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-update_logLik"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-update_logLik}{}}}
\subsection{Method \code{update_logLik()}}{
Update the log-likelihood inside object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$update_logLik()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-update_risk_drivers"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-update_risk_drivers}{}}}
\subsection{Method \code{update_risk_drivers()}}{
Update the clear sky and risk drivers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$update_risk_drivers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-update_NM_classification"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-update_NM_classification}{}}}
\subsection{Method \code{update_NM_classification()}}{
Update the classification of the Bernoulli random variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$update_NM_classification(filter = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filter}}{Logical, when \code{TRUE} before the classification will be runned the command
\code{ self$NM_model$filter()} to update the mixture classification.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-filter"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-filter}{}}}
\subsection{Method \code{filter()}}{
Filter the time series when new parameters are supplied in the method \verb{$update(params)}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$filter(fit = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit}}{Logical, when \code{TRUE}, if in the model's specification, the monthly mean and variances will be re estimated and the seasonal variance corrected
such that the total variance of the deseasonalized residuals is zero.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Update the slots \verb{$data}, \verb{$seasonal_data}, \verb{$monthly_data}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-Moments"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-Moments}{}}}
\subsection{Method \code{Moments()}}{
Compute the conditional moments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$Moments(t_now, t_hor, theta = 0, quiet = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_now}}{Character date. Today date.}

\item{\code{t_hor}}{Character date. Horizon date.}

\item{\code{theta}}{Numeric, shift parameter for the mixture.}

\item{\code{quiet}}{Logical for verbose messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-VaR"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-VaR}{}}}
\subsection{Method \code{VaR()}}{
Value at Risk for a \code{solarModel}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$VaR(moments, t_now, t_hor, theta = 0, ci = 0.05)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{moments}}{moments dataset}

\item{\code{t_now}}{Character date. Today date.}

\item{\code{t_hor}}{Character date. Horizon date.}

\item{\code{theta}}{Numeric, shift parameter for the mixture.}

\item{\code{ci}}{Confidence interval (one tail).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-logLik"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-logLik}{}}}
\subsection{Method \code{logLik()}}{
Compute the log-likelihood of the model and update the slot \verb{$loglik}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$logLik(moments, target = "Yt", quasi = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{moments}}{Dataset containing the moments to use for computation.}

\item{\code{target}}{Character. Target variable to use "Yt" or "GHI".}

\item{\code{quasi}}{Logical, when \code{TRUE} is computed the pseudo-likelihood with Gaussian link.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-print"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-print}{}}}
\subsection{Method \code{print()}}{
Print method for \code{solarModel} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
