% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-selection.R
\name{solarModel_AIC_BIC}
\alias{solarModel_AIC_BIC}
\title{Compute the AIC and BIC of a solarModel object}
\usage{
solarModel_AIC_BIC(model, target = "GHI", type = c("train", "test", "full"))
}
\arguments{
\item{model}{solarmodel}
}
\description{
Compute the AIC and BIC of a solarModel object
}
\note{
Version 1.0.0.
}
\keyword{solarModel}
