% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarMoments.R
\name{solarModel_calibrate_theta}
\alias{solarModel_calibrate_theta}
\title{Calibrate theta to match a certain level of solar radiation}
\usage{
solarModel_calibrate_theta(model, moments, e_RT_target, quiet = FALSE)
}
\arguments{
\item{e_RT_target}{Numeric, vector of target expectation to match.}
}
\description{
Calibrate theta to match a certain level of solar radiation
}
\details{
Version 1.0.0.
}
\keyword{solarModel}
