% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-internals.R
\name{solarModel_covariance}
\alias{solarModel_covariance}
\title{Compute the covariance}
\usage{
solarModel_covariance(
  t_now,
  mom_t,
  mom_T,
  GARCH,
  NM_model,
  theta = 0,
  tol = 0.01
)
}
\description{
Compute the covariance
}
\note{
Version 1.0.0.
}
\keyword{solarModel}
