% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-internals.R
\name{solarModel_forecast}
\alias{solarModel_forecast}
\title{Iterate the forecast on multiple dates}
\usage{
solarModel_forecast(model, moments, ci = 0.1, lambda = 0)
}
\description{
Iterate the forecast on multiple dates
}
\note{
Version 1.0.0.
}
\keyword{solarModel}
