% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-internals.R
\name{solarModel_predict}
\alias{solarModel_predict}
\title{Produce a forecast from a solarModel object}
\usage{
solarModel_predict(model, moments, lambda = 0, ci = 0.01)
}
\arguments{
\item{lambda}{Numeric scalar, Sugeno parameter.}
}
\description{
Produce a forecast from a solarModel object
}
\note{
Version 1.0.0.
}
\examples{
model = solarModel$new(spec)
model$fit()
moments <- model$moments$conditional[14,]
object <- solarModel_predict(model, moments, ci = 0.01)
object

}
\keyword{solarModel}
