% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{solarModel_predict_plot}
\alias{solarModel_predict_plot}
\title{Plot a forecast from a solarModel object}
\usage{
solarModel_predict_plot(object, pdf_components = FALSE, type = "mix")
}
\description{
Plot a forecast from a solarModel object
}
\examples{
model = solarModel$new(spec)
model$fit()
df_n <- model$moments$conditional[23,]
solarModel_predict_plot(solarModel_predict(model, df_n, ci = 0.01))
}
