% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-QMLE.R
\name{solarModel_QMLE}
\alias{solarModel_QMLE}
\title{QMLE Estimate}
\usage{
solarModel_QMLE(model, maxrestarts = 1, seed = 1, quiet = TRUE)
}
\description{
QMLE Estimate
}
\note{
Version 1.0.0.
}
\keyword{solarModel}
