% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-selection.R
\name{solarModel_selection}
\alias{solarModel_selection}
\alias{solarModels_selection}
\title{Select the Best Model}
\usage{
solarModels_selection(
  spec,
  arOrder = 2,
  maOrder = 2,
  archOrder = 1,
  garchOrder = 1
)
}
\arguments{
\item{spec}{specification}
}
\description{
Select the Best Model
}
\note{
Version 1.0.0.
}
\keyword{solarModel}
