% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_spec-R6.R
\name{solarModel_spec}
\alias{solarModel_spec}
\title{Control function for a \code{solarModel} object}
\description{
Control function for a \code{solarModel} object

Control function for a \code{solarModel} object
}
\details{
Control function for a \code{solarModel} object that contains all the setups used for the estimation.
}
\note{
Version 1.0.0.
}
\examples{
control <- solarModel_spec$new()
}
\keyword{solarModel}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{place}}{Character, optional name of the location considered.}

\item{\code{target}}{Character, name of the target variable to model. Can be \code{"GHI"} or \code{"clearsky"}.}

\item{\code{coords}}{A named list with the coordinates of the location considered. Contains:
\describe{
\item{lat}{Numeric, reference latitude in degrees.}
\item{lon}{Numeric, reference longitude in degrees.}
\item{alt}{Numeric, reference altitude in metres.}
}}

\item{\code{dates}}{A named list, with three sub-lists: \code{data} containing the information on the complete dataset,
\code{train} containing the information on the train dataset and \code{test} containing the information on the test dataset.
Each sub-list is structured as follows:
\describe{
\item{from}{Character date, minmum date in the dataset.}
\item{to}{Character date, maximum date in the dataset.}
\item{nobs}{Integer scalar, number of observations contained in the dataset between \code{from} and \code{to}.}
\item{perc}{Numeric scalar, percentage of data in the dataset with respect to the complete data.}
}}

\item{\code{data}}{Tibble, dataset with CAMS solar radiation data.}

\item{\code{transform}}{Named list, specification of the solar transform.}

\item{\code{clearsky}}{Named list, specification of the clear sky model.}

\item{\code{seasonal.mean}}{Named list, specification of the seasonal model.}

\item{\code{mean.model}}{Named list, specification of the ARMA model.}

\item{\code{seasonal.variance}}{Named list, specification of the seasonal variance model.}

\item{\code{variance.model}}{Named list, specification of the GARCH model for deseasonalized residuals \eqn{\tilde{e}_t}.}

\item{\code{mixture.model}}{Named list, specification of the Mixture model for GARCH residuals \eqn{u_t}.}

\item{\code{garch_variance}}{Logical, when \code{TRUE} the GARCH model will be used otherwise no.}

\item{\code{clearsky_threshold}}{Numeric, parameter > 1, used to scale up CAMS clearsky.}

\item{\code{stochastic_clearsky}}{Logical, when \code{TRUE} the clear sky is considered stochastic.}

\item{\code{quiet}}{Logical. When \code{TRUE} the function will not display any message. The dafault if \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-solarModel_spec-new}{\code{solarModel_spec$new()}}
\item \href{#method-solarModel_spec-specification}{\code{solarModel_spec$specification()}}
\item \href{#method-solarModel_spec-set_params}{\code{solarModel_spec$set_params()}}
\item \href{#method-solarModel_spec-set_transform}{\code{solarModel_spec$set_transform()}}
\item \href{#method-solarModel_spec-set_clearsky}{\code{solarModel_spec$set_clearsky()}}
\item \href{#method-solarModel_spec-set_seasonal.mean}{\code{solarModel_spec$set_seasonal.mean()}}
\item \href{#method-solarModel_spec-set_mean.model}{\code{solarModel_spec$set_mean.model()}}
\item \href{#method-solarModel_spec-set_seasonal.variance}{\code{solarModel_spec$set_seasonal.variance()}}
\item \href{#method-solarModel_spec-set_variance.model}{\code{solarModel_spec$set_variance.model()}}
\item \href{#method-solarModel_spec-set_mixture.model}{\code{solarModel_spec$set_mixture.model()}}
\item \href{#method-solarModel_spec-print}{\code{solarModel_spec$print()}}
\item \href{#method-solarModel_spec-clone}{\code{solarModel_spec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel_spec-new"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel_spec-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{solarModel_spec} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel_spec$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel_spec-specification"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel_spec-specification}{}}}
\subsection{Method \code{specification()}}{
Specification function for a \code{solarModel}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel_spec$specification(
  place,
  target = "GHI",
  min_date,
  max_date,
  from,
  to,
  data
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{place}}{Character, name of an element in the \code{CAMS_data} list.}

\item{\code{target}}{Character, target variable to model. Can be \code{GHI} or \code{clearsky}.}

\item{\code{min_date}}{Character. Date in the format \code{YYYY-MM-DD}. Minimum date for the complete data. If \code{missing} will be used the minimum data available.}

\item{\code{max_date}}{Character. Date in the format \code{YYYY-MM-DD}. Maximum date for the complete data. If \code{missing} will be used the maximum data available.}

\item{\code{from}}{Character. Date in the format \code{YYYY-MM-DD}. Starting date to use for training data.
If \code{missing} will be used the minimum data available after filtering for \code{min_date}.}

\item{\code{to}}{character. Date in the format \code{YYYY-MM-DD}. Ending date to use for training data.
If \code{missing} will be used the maximum data available after filtering for \code{max_date}.}

\item{\code{data}}{data for the selected location.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel_spec-set_params"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel_spec-set_params}{}}}
\subsection{Method \code{set_params()}}{
Generic controls
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel_spec$set_params(
  stochastic_clearsky = FALSE,
  clearsky_threshold = 1.01,
  quiet = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stochastic_clearsky}}{Logical, when \code{TRUE} the clear sky will be considered stochastic.}

\item{\code{clearsky_threshold}}{Numeric, parameter > 1, used to scale up CAMS clearsky to avoid that clear sky radiaion and global horizontal radiation are equal.}

\item{\code{quiet}}{Logical. When \code{TRUE} the function will not display any message. The dafault if \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel_spec-set_transform"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel_spec-set_transform}{}}}
\subsection{Method \code{set_transform()}}{
Control parameters for the \code{solarTransform}. See \code{\link{solarTransform}} for more details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel_spec$set_transform(
  min_pos = 1,
  max_pos = 1,
  link = "invgumbel",
  delta = 0.05,
  threshold = 0.01
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min_pos}}{Integer, position of the minimum. For example when \code{2} the minimum is the second lowest value.}

\item{\code{max_pos}}{Integer, position of the maximum. For example when \code{3} the maximum is the third greatest value.}

\item{\code{link}}{Character, link function.}

\item{\code{delta}}{transform params}

\item{\code{threshold}}{Numeric. Threshold used to estimate the transformation parameters \deqn{\alpha} and \deqn{\beta}.
The default is \code{0.01}. See \code{\link{solarTransform}} for more details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel_spec-set_clearsky"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel_spec-set_clearsky}{}}}
\subsection{Method \code{set_clearsky()}}{
List with specification's parameters of the clear sky model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel_spec$set_clearsky(control = control_seasonalClearsky())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{control}}{Named list with control parameters. See \code{\link{control_seasonalClearsky}} for more details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel_spec-set_seasonal.mean"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel_spec-set_seasonal.mean}{}}}
\subsection{Method \code{set_seasonal.mean()}}{
List with specification's parameters of the seasonal mean \eqn{\bar{Y}_t} for \eqn{Y_t}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel_spec$set_seasonal.mean(
  order = 1,
  period = 365,
  include.trend = FALSE,
  include.intercept = TRUE,
  monthly.mean = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{order}}{Integer. Specify the order of the seasonal mean \deqn{\bar{Y}_t}. The default is \code{1}.}

\item{\code{period}}{Integer, seasonal periodicity, the default is \code{365}.}

\item{\code{include.trend}}{Logical. When \code{TRUE} an yearly trend \deqn{t} will be included in the seasonal model, otherwise will be excluded. The default is \code{FALSE}.}

\item{\code{include.intercept}}{Logical. When \code{TRUE} the intercept \deqn{a_0} will be included in the seasonal model, otherwise will be excluded. The default is \code{TRUE}.}

\item{\code{monthly.mean}}{Logical. When \code{TRUE} a vector of 12 monthly means will be computed on the deseasonalized series \deqn{\tilde{Y}_t = Y_t - \bar{Y}_t}
and it is subtracted to ensure that the time series is centered around zero for all the months. The dafault if \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel_spec-set_mean.model"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel_spec-set_mean.model}{}}}
\subsection{Method \code{set_mean.model()}}{
List with specification's parameters of the ARMA model for deseasonalized series \eqn{\tilde{Y}_t = Y_t - \bar{Y}_t}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel_spec$set_mean.model(
  arOrder = 1,
  maOrder = 0,
  include.intercept = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{arOrder}}{Integer. An integer specifying the order of the AR component. The default is \code{1}.}

\item{\code{maOrder}}{Integer. An integer specifying the order of the MA component. The default is \code{0}.}

\item{\code{include.intercept}}{Logical. When \code{TRUE} the intercept \deqn{\phi_0} will be included in the seasonal model, otherwise will be excluded. The default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel_spec-set_seasonal.variance"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel_spec-set_seasonal.variance}{}}}
\subsection{Method \code{set_seasonal.variance()}}{
List with specification's parameters of the seasonal variance \eqn{\bar{\sigma}_t} for ARMA's residuals \eqn{e_t}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel_spec$set_seasonal.variance(
  order = 1,
  period = 365,
  include.trend = FALSE,
  correction = TRUE,
  monthly.mean = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{order}}{Integer. Specify the order of the seasonality of the seasonal variance. The default is \code{1}.}

\item{\code{period}}{Integer, seasonal periodicity, the default is \code{365}.}

\item{\code{include.trend}}{Logical. When \code{TRUE} an yearly trend \deqn{t} will be included in the seasonal model, otherwise will be excluded. The default is \code{FALSE}.}

\item{\code{correction}}{Logical. When \code{TRUE} the parameters of seasonal variance are corrected to ensure
that the standardize the residuals have exactly a unitary variance. The dafault if \code{TRUE}.}

\item{\code{monthly.mean}}{Logical. When \code{TRUE} a vector of 12 monthly std. deviations will be computed
on the standardized residuals  \deqn{\tilde{\varepsilon}_t} and used to standardize the time series
such that it has unitary variance for all the months. The default if \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel_spec-set_variance.model"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel_spec-set_variance.model}{}}}
\subsection{Method \code{set_variance.model()}}{
List with specification's parameters of the GARCH variance \eqn{\sigma_t} for deseasonalized residuals \eqn{\tilde{e}_t = e_t/\bar{\sigma}_t}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel_spec$set_variance.model(
  archOrder = 1,
  garchOrder = 1,
  garch_variance = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{archOrder}}{Integer. An integer specifying the order of the ARCH component. The default is \code{1}.}

\item{\code{garchOrder}}{Integer. An integer specifying the order of the GARCH component. The default is \code{1}.}

\item{\code{garch_variance}}{Logical. When \code{TRUE} the GARCH model will be used to standardize the residuals otherwise will be excluded. The dafault if \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel_spec-set_mixture.model"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel_spec-set_mixture.model}{}}}
\subsection{Method \code{set_mixture.model()}}{
List with specification's parameters of the Gaussian mixture model for GARCH residuals \eqn{u_t = \tilde{e}_t/\sigma_t}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel_spec$set_mixture.model(
  abstol = 1e-20,
  match.expectation = TRUE,
  match.variance = FALSE,
  match.empiric = FALSE,
  method = "mclust",
  maxit = 5000,
  maxrestarts = 500
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{abstol}}{Numeric. Absolute level for convergence of the EM-algorithm. The default is \code{1e-20}.}

\item{\code{match.expectation}}{Logical, when \code{TRUE} the mixture parameters ensures that the expected value is matched.}

\item{\code{match.variance}}{Logical, when \code{TRUE} the mixture parameters ensures that the variance is matched.}

\item{\code{match.empiric}}{Logical, when \code{TRUE} and \code{match.expectation = TRUE} or  \code{match.variance = TRUE} the mixture parameters
will be estimated ensuring that mean and variance matches the empirical parameters. Otherwise if \code{FALSE} and
\code{match.expectation = TRUE} or \code{match.variance = TRUE} the target expectation will be zero and the target variance 1.}

\item{\code{method}}{Character, package used to fit the parameters. Can be \code{mclust} or \code{mixtools}.}

\item{\code{maxit}}{Integer. Maximum number of iterations for EM-algorithm. The default is \code{5000}.}

\item{\code{maxrestarts}}{Integer. Maximum number of restarts when EM-algorithm does not converge. The default is \code{500}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel_spec-print"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel_spec-print}{}}}
\subsection{Method \code{print()}}{
Print method for \code{solarModel_spec} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel_spec$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModel_spec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-solarModel_spec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModel_spec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
