% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-diagnostics.R
\name{solarModel_test_autocorr}
\alias{solarModel_test_autocorr}
\title{Autocorrelation test}
\usage{
solarModel_test_autocorr(
  model,
  lag.max = 3,
  ci = 0.05,
  method = c("bg", "bp", "lb"),
  type = c("train", "test", "full")
)
}
\arguments{
\item{model}{An object of the class \code{solarModel}}

\item{lag.max}{Numeric, scalar. Maximum lag to consider for the test.}

\item{ci}{Numeric, scalar. Minimum p-value to consider the test \code{"passed"}.}

\item{method}{Character, type of test. Can be \code{"bg"} for Breusch-Godfrey, \code{"bp"} for Box-pierce and \code{"lb"} for BLjung-Box.}

\item{train_data}{Logical, when \code{TRUE} only train data will be used to evaluate the test, otherwise all the available sample.}
}
\description{
Evaluate the autocorrelation in the components of a \code{solarModel} object.
}
\note{
Version 1.0.0.
}
\examples{
model = solarModel$new(spec)
model$fit()
solarModel_test_autocorr(model, method = "lb")

}
\keyword{solarModel_test}
