% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-diagnostics.R
\name{solarModel_test_distribution}
\alias{solarModel_test_distribution}
\title{Distribution test}
\usage{
solarModel_test_distribution(
  model,
  H0 = c("gm", "norm"),
  ci = 0.05,
  min_quantile = 0.025,
  max_quantile = 0.985,
  type = c("train", "test", "full")
)
}
\arguments{
\item{model}{An object of the class \code{solarModel}}

\item{ci}{p.value for rejection.}

\item{min_quantile}{minimum quantile for the grid of values.}

\item{max_quantile}{maximum quantile for the grid of values.}

\item{type}{Type of test.}

\item{test}{Character, null hypothesis for the residuals distribution \code{"gm"} for Gaussian mixture and \code{"norm"} for normality.}
}
\description{
Evaluate a Kolmogorov-Smirnov test on the residuals of a \code{solarModel} model
object against the estimated Gaussian mixture distribution.
}
\note{
Version 1.0.0.
}
\examples{
model = solarModel$new(spec)
model$fit()
solarModel_test_distribution(model)

}
\keyword{solarModel_test}
