% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-diagnostics.R
\name{solarModel_test_forecast}
\alias{solarModel_test_forecast}
\title{Compute metrics to test forecasts}
\usage{
solarModel_test_forecast(model, ci = 0.1, type = c("train", "test", "full"))
}
\description{
Compute metrics to test forecasts
}
\note{
Version 1.0.0.
}
\keyword{solarModel_test}
