% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-diagnostics.R
\name{solarModel_test_LPD}
\alias{solarModel_test_LPD}
\title{Compute the Log-predictive density of a solarModel}
\usage{
solarModel_test_LPD(model, type = c("train", "test", "full"))
}
\description{
Compute the Log-predictive density of a solarModel
}
\note{
Version 1.0.0.
}
\keyword{solarModel_test}
