% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-diagnostics.R
\name{solarModel_test_PIT}
\alias{solarModel_test_PIT}
\title{Probability Integral Transform}
\usage{
solarModel_test_PIT(model, ci = 0.05, type = c("train", "test", "full"))
}
\description{
Probability Integral Transform
}
\note{
Version 1.0.0.
}
\keyword{solarModel_test}
