% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-diagnostics.R
\name{solarModel_test_pricing}
\alias{solarModel_test_pricing}
\title{Compute metrics to test option pricing}
\usage{
solarModel_test_pricing(
  model,
  type = c("train", "test", "full"),
  control = control_solarOption()
)
}
\description{
Compute metrics to test option pricing
}
\note{
Version 1.0.0.
}
\keyword{solarModel_test}
