% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-diagnostics.R
\name{solarModel_tests}
\alias{solarModel_tests}
\title{Autocorrelation and distribution tests}
\usage{
solarModel_tests(
  model,
  lags = c(7),
  ci = 0.05,
  min_quantile = 0.025,
  max_quantile = 0.985,
  method = "bg",
  type = c("train", "test", "full")
)
}
\arguments{
\item{model}{An object of the class \code{solarModel}}

\item{lags}{Numeric vector. Lags on which perform the autocorrelation tests. Can be more than one.}

\item{ci}{p.value for rejection.}

\item{min_quantile}{minimum quantile for the grid of values.}

\item{max_quantile}{maximum quantile for the grid of values.}

\item{method}{Character, type of test. Can be \code{"bg"} for Breusch-Godfrey, \code{"bp"} for Box-pierce and \code{"lb"} for BLjung-Box.}

\item{type}{Type of test.}
}
\description{
Evaluate a Kolmogorov-Smirnov test on the residuals of a \code{solarModel} model
object against the estimated Gaussian mixture distribution and a Breush-pagan or Box-pierce
test on the residuals.
}
\note{
Version 1.0.0.
}
\examples{
model = solarModel$new(spec)
model$fit()
solarModel_tests(model, train_data = TRUE)
}
\keyword{solarModel_test}
