% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-VaR.R
\name{solarModel_VaR}
\alias{solarModel_VaR}
\title{Compute the Value at Risk and Expected Shortfall of a SolarModel}
\usage{
solarModel_VaR(model, alpha = 0.05, ci = 0.05, ES = FALSE, type = "full")
}
\arguments{
\item{model}{solarModel}

\item{alpha}{Numeric vector of confidence levels. Allows for more than one alpha.}

\item{ci}{Numeric scalar, confidence levels used to state if the Null is rejected or not on VaR tests.}

\item{ES}{Logical, when \code{TRUE} the expected shortfall will be also computed for each alpha.}

\item{type}{Numeric, type of evaluation, \code{full} on the complete data, \code{train} on the train data, \code{test} on the test data.}
}
\description{
Compute the Value at Risk and Expected Shortfall of a SolarModel
}
