% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-selection.R
\name{solarModels_grid}
\alias{solarModels_grid}
\title{Fit a grid of solarModels}
\usage{
solarModels_grid(
  spec,
  arOrder = 2,
  maOrder = 2,
  archOrder = 1,
  garchOrder = 1,
  QMLE = FALSE
)
}
\arguments{
\item{spec}{Specification}

\item{arOrder}{Numeric, maximum AR order.}

\item{maOrder}{Numeric, maximum MA order.}

\item{archOrder}{Numeric, maximum ARCH order.}

\item{garchOrder}{Numeric, maximum GARCH order.}

\item{place}{Reference location}
}
\description{
Fit a grid of solarModels
}
\note{
Version 1.0.0.
}
\examples{
spec <- solarModel_spec$new()
models <- solarModels_grid(spec, "Bologna", 1,1,1,1)
models[which.min(models$L),]

}
\keyword{solarModel}
