% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarMoment-R6.R
\name{solarMoment}
\alias{solarMoment}
\title{Compute the generic conditional moments of a solarModel object}
\description{
Compute the generic conditional moments of a solarModel object

Compute the generic conditional moments of a solarModel object
}
\details{
Version 1.0.0.
}
\keyword{solarMoments}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-solarMoment-new}{\code{solarMoment$new()}}
\item \href{#method-solarMoment-filter}{\code{solarMoment$filter()}}
\item \href{#method-solarMoment-filter_ARMA}{\code{solarMoment$filter_ARMA()}}
\item \href{#method-solarMoment-filter_NM}{\code{solarMoment$filter_NM()}}
\item \href{#method-solarMoment-filter_GARCH}{\code{solarMoment$filter_GARCH()}}
\item \href{#method-solarMoment-filter_weights}{\code{solarMoment$filter_weights()}}
\item \href{#method-solarMoment-pdf_Y}{\code{solarMoment$pdf_Y()}}
\item \href{#method-solarMoment-cdf_Y}{\code{solarMoment$cdf_Y()}}
\item \href{#method-solarMoment-pdf_R}{\code{solarMoment$pdf_R()}}
\item \href{#method-solarMoment-cdf_R}{\code{solarMoment$cdf_R()}}
\item \href{#method-solarMoment-Q_R}{\code{solarMoment$Q_R()}}
\item \href{#method-solarMoment-print}{\code{solarMoment$print()}}
\item \href{#method-solarMoment-clone}{\code{solarMoment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-new"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$new(model, t_now, t_hor)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-filter"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-filter}{}}}
\subsection{Method \code{filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$filter(theta = 0, B, t_cond)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-filter_ARMA"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-filter_ARMA}{}}}
\subsection{Method \code{filter_ARMA()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$filter_ARMA()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-filter_NM"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-filter_NM}{}}}
\subsection{Method \code{filter_NM()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$filter_NM(theta = 0, B, t_cond)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-filter_GARCH"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-filter_GARCH}{}}}
\subsection{Method \code{filter_GARCH()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$filter_GARCH()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-filter_weights"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-filter_weights}{}}}
\subsection{Method \code{filter_weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$filter_weights()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-pdf_Y"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-pdf_Y}{}}}
\subsection{Method \code{pdf_Y()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$pdf_Y(type = c("mix", "up", "dw"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-cdf_Y"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-cdf_Y}{}}}
\subsection{Method \code{cdf_Y()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$cdf_Y(type = c("mix", "up", "dw"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-pdf_R"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-pdf_R}{}}}
\subsection{Method \code{pdf_R()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$pdf_R(type = c("mix", "up", "dw"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-cdf_R"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-cdf_R}{}}}
\subsection{Method \code{cdf_R()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$cdf_R(type = c("mix", "up", "dw"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-Q_R"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-Q_R}{}}}
\subsection{Method \code{Q_R()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$Q_R(type = c("mix", "up", "dw"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-print"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
