% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarMoments.R
\name{solarMoments}
\alias{solarMoments}
\title{Compute the generic conditional moments of a solarModel object}
\usage{
solarMoments(
  t_now,
  t_hor,
  data,
  ARMA,
  GARCH,
  NM_model,
  transform,
  theta = 0,
  quiet = FALSE
)
}
\arguments{
\item{t_now}{Date for today.}

\item{t_hor}{Horizon date.}

\item{data}{Slot \code{data} of a \code{solarModel} object. See the function \code{\link{solarModel}} for details.}

\item{ARMA}{Slot \code{ARMA} of a \code{solarModel} object.}

\item{GARCH}{Slot \code{GARCH} of a \code{solarModel} object.}

\item{NM_model}{Slot \code{NM_model} of a \code{solarModel} object.}

\item{transform}{Slot \code{transform} of a \code{solarModel} object.}

\item{theta}{Numeric, vector of shift parameters for the Gaussian mixture residuals.}
}
\description{
Compute the generic conditional moments of a solarModel object
}
\details{
Version 1.0.0.
}
\examples{
model = solarModel$new(spec)
model$fit()
t_now = "2019-07-11"
t_hor = "2019-10-19"
data = model$data
ARMA = model$ARMA
GARCH  = model$GARCH
NM_model = model$NM_model
transform = model$transform
theta = 0
solarMoments(t_now, t_hor, data, ARMA, GARCH, NM_model, transform, theta = 0, quiet = FALSE)
filter(model$moments$conditional, date == t_hor)
filter(model$moments$unconditional, date == t_hor)

t_seq <- seq.Date(as.Date("2013-01-01"), as.Date("2013-12-31"), 1)
mom <- purrr::map_df(t_seq, ~solarr::solarMoments(.x-150, .x, data, ARMA, GARCH, NM_model, transform, theta = 0, quiet = FALSE))
solarOption_model(model, mom, control_options = control_solarOption(nyears = c(2005, 2024)))
solarOption_historical(model, control_options = control_solarOption(nyears = c(2005, 2024)))

}
\keyword{solarMoments}
