% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarMoments.R
\name{solarMoments_conditional}
\alias{solarMoments_conditional}
\title{Compute the conditional moments}
\usage{
solarMoments_conditional(data, theta = 0, control_model)
}
\arguments{
\item{data}{Slot \code{data} of a \code{solarModel} object. See the function \code{\link{solarModel}} for details.}

\item{theta}{Numeric, shift parameter for the Gaussian mixture residuals.}

\item{control_model}{An object with the class \code{solarModel_spec}. See the function \code{\link{solarModel_spec}} for details.}
}
\description{
Compute the conditional moments
}
\details{
Version 1.0.0.
}
\keyword{solarMoments}
