% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarMoments.R
\name{solarMoments_path}
\alias{solarMoments_path}
\title{Condition the moments for a specific Bernoulli realization at a time t_cond}
\usage{
solarMoments_path(moments, GARCH, NM_model, theta = 0, B = 1, t_cond)
}
\arguments{
\item{GARCH}{Slot \code{GARCH} of a \code{solarModel} object.}

\item{NM_model}{Slot \code{NM_model} of a \code{solarModel} object.}

\item{theta}{Numeric, vector of shift parameters for the Gaussian mixture residuals.}

\item{B}{conditioning value for the Bernoulli state at time thor}

\item{t_cond}{conditioning date}
}
\description{
Condition the moments for a specific Bernoulli realization at a time t_cond
}
\details{
Version 1.0.0.
}
\examples{
model = solarModel$new(spec)
model$fit()
GARCH  = model$GARCH
NM_model = model$NM_model
# Compute the moments
moments <- model$Moments("2012-01-01", "2012-03-05")

# Condition the moments on a realization of B
moments
solarOption_model(model, moments)

mom_cond <- solarMoments_path(moments, GARCH, NM_model, theta = 0, B = 1, "2012-02-03")
solarOption_model(model, mom_cond)

mom_cond <- solarMoments_path(moments, GARCH, NM_model, theta = 0, B = 1, "2012-03-01")
solarOption_model(model, mom_cond)

mom_cond <- solarMoments_path(moments, GARCH, NM_model, theta = 0, B = 1, "2012-03-04")
solarOption_model(model, mom_cond)

mom_cond <- solarMoments_path(moments, GARCH, NM_model, theta = 0, B = 0, "2012-03-04")
solarOption_model(model, mom_cond)

mom_cond <- solarMoments_path(moments, GARCH, NM_model, theta = 0, B = c(0, 0), c("2012-03-03", "2012-03-04"))
solarOption_model(model, mom_cond)

mom_cond <- solarMoments_path(moments, GARCH, NM_model, theta = 0, B = c(1, 1), c("2012-03-03", "2012-03-04"))
solarOption_model(model, mom_cond)

}
\keyword{solarMoments}
