% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarMoments.R
\name{solarMoments_unconditional}
\alias{solarMoments_unconditional}
\title{Compute the unconditional moments}
\usage{
solarMoments_unconditional(data, ARMA, GARCH, theta = 0)
}
\arguments{
\item{data}{Slot \code{data} of a \code{solarModel} object. See the function \code{\link{solarModel}} for details.}

\item{ARMA}{Slot \code{ARMA} of a \code{solarModel} object.}

\item{GARCH}{Slot \code{GARCH} of a \code{solarModel} object.}

\item{theta}{Numeric, shift parameter for the Gaussian mixture residuals.}
}
\description{
Compute the unconditional moments
}
\details{
Version 1.0.0.
}
\keyword{solarMoments}
