% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption-R6.R
\name{solarOption}
\alias{solarOption}
\title{Create a SoRad / SoREd contract specification}
\description{
Create a SoRad / SoREd contract specification

Create a SoRad / SoREd contract specification
}
\note{
Version 1.0.0.
}
\keyword{solarOption}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ticker}}{description}

\item{\code{strike}}{Strike price for solar radiation.}

\item{\code{t_pricing}}{Character, pricing date.}

\item{\code{t_now}}{Character, today date.}

\item{\code{t_init}}{Character, inception date.}

\item{\code{t_hor}}{Character, maturity date.}

\item{\code{tick}}{Numeric, monetary conversion tick.}

\item{\code{contract_type}}{Character, maturity date.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{control}}{control parameters}

\item{\code{tau}}{Numeric, scalar. Time from \code{t_now} till \code{t_hor} in days.}

\item{\code{tau_accrued}}{Numeric, scalar. Time from \code{t_pricing} till \code{t_hor} in days.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-solarOption-new}{\code{solarOption$new()}}
\item \href{#method-solarOption-set_contract}{\code{solarOption$set_contract()}}
\item \href{#method-solarOption-set_control}{\code{solarOption$set_control()}}
\item \href{#method-solarOption-print}{\code{solarOption$print()}}
\item \href{#method-solarOption-clone}{\code{solarOption$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarOption-new"></a>}}
\if{latex}{\out{\hypertarget{method-solarOption-new}{}}}
\subsection{Method \code{new()}}{
Initialize the contract
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarOption$new(contract_type = "SoRad")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{contract_type}}{Character, contract type "SoRad" or "SoREd"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarOption-set_contract"></a>}}
\if{latex}{\out{\hypertarget{method-solarOption-set_contract}{}}}
\subsection{Method \code{set_contract()}}{
Initialize the contract
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarOption$set_contract(t_pricing, t_init, t_hor, strike, tick = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_pricing}}{Character, pricing date.}

\item{\code{t_init}}{Character, inception date.}

\item{\code{t_hor}}{Character, maturity date.}

\item{\code{strike}}{Numeric, strike price.}

\item{\code{tick}}{Numeric monetary tick.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarOption-set_control"></a>}}
\if{latex}{\out{\hypertarget{method-solarOption-set_control}{}}}
\subsection{Method \code{set_control()}}{
Store a list of custom control parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarOption$set_control(control)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{control}}{List, control parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarOption-print"></a>}}
\if{latex}{\out{\hypertarget{method-solarOption-print}{}}}
\subsection{Method \code{print()}}{
Print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarOption$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarOption-clone"></a>}}
\if{latex}{\out{\hypertarget{method-solarOption-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarOption$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
