% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarMoments.R
\name{solarOption_calibrate_theta}
\alias{solarOption_calibrate_theta}
\title{Calibrate the time series of theta to match a certain level of Option price}
\usage{
solarOption_calibrate_theta(
  model,
  moments,
  P_target,
  put = TRUE,
  quiet = FALSE,
  control_options = control_solarOption()
)
}
\arguments{
\item{model}{An object with the class \code{solarModel}. See the function \code{\link{solarModel}} for details.}

\item{P_target}{Numeric, vector of target prices to match.}

\item{put}{Logical, when \code{TRUE}, the default, will be computed the price for a \code{put} contract, otherwise for a \code{call} contract.}

\item{control_options}{Named list with control parameters. See \code{\link{control_solarOption}} for more details.}
}
\description{
Calibrate the time series of theta to match a certain level of Option price
}
\details{
Version 1.0.0.
}
\keyword{solarOption}
