% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_choquet}
\alias{solarOption_choquet}
\title{Compute Choquet price for a Solar Option}
\usage{
solarOption_choquet(
  model,
  moments,
  portfolio,
  nmonths = 1:12,
  lambda = 0,
  implvol = 1,
  put = TRUE,
  control_options = control_solarOption()
)
}
\arguments{
\item{model}{An object with the class \code{solarModel}. See the function \code{\link{solarModel}} for details.}

\item{moments}{Tibble containing the forecasted moments for different days ahead. See the function \code{\link{solarMoments}} for more details.}

\item{portfolio}{Optional, A list of objects of the class \code{solarOptionPortfolio}.}

\item{nmonths}{Numeric vector, months in which the payoff should be computed. Can vary from 1 (January) to 12 (December).}

\item{lambda}{Numeric, Sugeno parameter.}

\item{implvol}{Numeric, implied volatility.}

\item{put}{Logical, when \code{TRUE}, the default, will be computed the price for a \code{put} contract, otherwise for a \code{call} contract.}

\item{control_options}{Named list with control parameters. See \code{\link{control_solarOption}} for more details.}
}
\description{
Compute Choquet price for a Solar Option
}
\note{
Version 1.0.0.
}
\examples{
model = solarModel$new(spec)
model$fit()
moments <- model$moments$unconditional[1:365,]
lambda = 0
control_options = control_solarOption()
solarOption_choquet(model, moments[1:30,], lambda = 0.01)
solarOption_model(model, moments[1,])
solarOption_choquet(model, moments, lambda = 0.1, put = F)
}
\keyword{solarOption}
