% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_greeks}
\alias{solarOption_greeks}
\title{Compute the Greeks of a Solar Option}
\usage{
solarOption_greeks(model, moments, put = TRUE, elasticities = FALSE)
}
\description{
Compute the Greeks of a Solar Option
}
\note{
Version 1.0.0.
}
\examples{
# Model fit
model <- solarModel$new(spec)
model$fit()
solarOption_greeks(model, model$Moments("2023-01-01", "2024-05-01"))

}
\keyword{solarOption}
