% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_historical}
\alias{solarOption_historical}
\title{Payoff of solar options on historical data}
\usage{
solarOption_historical(
  model,
  nmonths = 1:12,
  put = TRUE,
  control_options = control_solarOption()
)
}
\arguments{
\item{model}{An object with the class \code{solarModel}. See the function \code{\link{solarModel}} for details.}

\item{nmonths}{Numeric vector, months in which the payoff should be computed. Can vary from 1 (January) to 12 (December).}

\item{put}{Logical, when \code{TRUE}, the default, will be computed the price for a \code{put} contract, otherwise for a \code{call} contract.}

\item{control_options}{Named list with control parameters. See \code{\link{control_solarOption}} for more details.}
}
\value{
An object of the class \code{solarOptionPayoff}.
}
\description{
Payoff of solar options on historical data
}
\note{
Version 1.0.0.
}
\examples{
# Solar model
model <- solarModel$new(spec)
model$fit()
# Historical payoff (put)
solarOption_historical(model, put=TRUE)
# Historical payoff (call)
solarOption_historical(model, put=FALSE)

}
\keyword{solarOption}
