% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_index_greeks}
\alias{solarOption_index_greeks}
\title{Compute the Greeks of a Solar Option index}
\usage{
solarOption_index_greeks(model, moments, elasticities = FALSE)
}
\description{
Compute the Greeks of a Solar Option index
}
\note{
Version 1.0.0.
}
\examples{
# Model fit
model <- solarModel$new(spec)
model$fit()

t_now <- as.Date("2024-01-01")
t_hor <- as.Date("2024-12-31")
dates <- seq.Date(t_now+1, t_hor, 1)
mom <- purrr::map_df(dates, ~model$Moments(t_now, .x))
solarOption_index_greeks(model, mom, elasticities = TRUE)
mom_test <- mom
mom_test$beta <- mom_test$beta*1.01
solarOption_index_greeks(model, mom_test, elasticities = TRUE)

}
\keyword{solarOption}
