% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption-internals.R
\name{solarOption_lambda}
\alias{solarOption_lambda}
\title{Calibrate the implied Choquet parameter}
\usage{
solarOption_lambda(
  model,
  moments,
  P_target,
  r_imp = 0,
  put = TRUE,
  quiet = FALSE,
  control_options = control_solarOption()
)
}
\arguments{
\item{model}{An object with the class \code{solarModel}. See the function \code{\link{solarModel}} for details.}

\item{moments}{Tibble containing the forecasted moments for different days ahead. See the function \code{\link{solarMoments}} for more details.}

\item{P_target}{Numeric, target price to match.}

\item{r_imp}{Numeric, implied return from measure under P.}

\item{put}{Logical, when \code{TRUE}, the default, will be computed the price for a \code{put} contract, otherwise for a \code{call} contract.}

\item{control_options}{Named list with control parameters. See \code{\link{control_solarOption}} for more details.}
}
\description{
Calibrate the implied Choquet parameter
}
\details{
The optimization function will find the best parameter \eqn{\lambda} such that
\deqn{\underset{-0.5 < \lambda < 1}{\text{argmin}} \left(P^{\mathbb{P}(\lambda)} (1 + r^{\text{imp}}) - P_{\text{target}}\right)^2}
}
\note{
Version 1.0.0.
}
\keyword{solarOption}
