% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_model}
\alias{solarOption_model}
\title{Compute the premium given the moments}
\usage{
solarOption_model(
  model,
  moments,
  portfolio,
  nmonths = 1:12,
  lambda = 0,
  implvol = 1,
  put = TRUE,
  control_options = control_solarOption()
)
}
\arguments{
\item{model}{An object with the class \code{solarModel}. See the function \code{\link{solarModel}} for details.}

\item{moments}{Tibble containing the forecasted moments for different days ahead. See the function \code{\link{solarMoments}} for more details.}

\item{portfolio}{Optional, A list of objects of the class \code{solarOptionPortfolio}.}

\item{nmonths}{Numeric vector, months in which the payoff should be computed. Can vary from 1 (January) to 12 (December).}

\item{lambda}{Numeric, Sugeno parameter.}

\item{implvol}{Numeric, implied volatility.}

\item{put}{Logical, when \code{TRUE}, the default, will be computed the price for a \code{put} contract, otherwise for a \code{call} contract.}

\item{control_options}{Named list with control parameters. See \code{\link{control_solarOption}} for more details.}
}
\description{
Compute the premium given the moments
}
\note{
Version 1.0.0.
}
\examples{
# Model
model = solarModel$new(spec)
model$fit()
# Pricing without portfolio
moments <- model$moments$unconditional
# Premium
premium_Vt <- solarOption_model(model, moments, put = TRUE)
# Pricing date
t_now <- as.Date("2021-12-31")
# Inception date
t_init <- as.Date("2022-01-01")
# Maturity date
t_hor <- as.Date("2022-12-31")
# solar options portfolio
portfolio <- SoRadPorfolio(model, t_now, t_init, t_hor)
# Moments
moments <- purrr::map_df(portfolio, ~model$Moments(t_now, .x$t_hor))
# Premium
premium_Vt <- solarOption_model(model, moments, portfolio, put = TRUE)
premium_Vt$payoff_year$premium
}
\keyword{solarOption}
