% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_moments}
\alias{solarOption_moments}
\title{Compute the first fourth moments and variance of a SoRad}
\usage{
solarOption_moments(
  moments,
  transform,
  lambda = 0,
  put = TRUE,
  link = "invgumbel",
  control_options = control_solarOption()
)
}
\arguments{
\item{transform}{slot transform of a solarModel}

\item{lambda}{Sugeno parameter}

\item{put}{Logical, when \code{TRUE}, the default, will be computed the price for a \code{put} contract, otherwise for a \code{call} contract.}

\item{control_options}{Named list with control parameters. See \code{\link{control_solarOption}} for more details.}
}
\description{
Compute the first fourth moments and variance of a SoRad
}
\note{
Version 1.0.0.
}
\examples{
# Solar model
model <- solarModel$new(spec)
model$fit()
moments <- model$moments$unconditional
solarOption_moments(moments[1:365,], model$transform)

}
\keyword{solarOption}
