% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption-internals.R
\name{solarOptionPayoff}
\alias{solarOptionPayoff}
\title{Structure the outputs of solarOption functions}
\usage{
solarOptionPayoff(data, leap_year = FALSE)
}
\arguments{
\item{data}{df_payoff}

\item{leap_year}{control params}
}
\value{
An object of the class \code{solarOptionPayoff}.
}
\description{
Structure the outputs of solarOption functions
}
\note{
Version 1.0.0.
}
\keyword{solarOption}
