% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption-internals.R
\name{solarPayoff}
\alias{solarPayoff}
\title{Payoff function of a Solar Option}
\usage{
solarPayoff(R, K = 0, put = TRUE)
}
\arguments{
\item{R}{Numeric, vector of values of solar radiation at maturity.}

\item{K}{Numeric, scalar or vector of strikes.}

\item{put}{Logical, when \code{TRUE}, the default, the function will return the output of a put payoff
otherwise a call payoff. See the details.}
}
\description{
Payoff function of a Solar Option
}
\details{
The put option payoff reads:
\deqn{(K - R)^{+} = (K - R) 1_{K > R}}
Symmetrically a call option payoff reads:
\deqn{(R-K)^{+} = (R - K) 1_{R \ge K}}
}
\note{
Version 1.0.0.
}
\examples{
solarPayoff(10, 9, put = TRUE)

}
\keyword{solarOption}
