% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{solarScenario_plot}
\alias{solarScenario_plot}
\title{Plot scenarios from a solarScenario object}
\usage{
solarScenario_plot(x, target = "GHI", nsim = 10, empiric = TRUE, ci = 0.05)
}
\description{
Plot scenarios from a solarScenario object
}
\examples{
model = solarModel$new(spec)
model$fit()
scenario <- solarScenario(model, nsim = 70)
solarScenario_plot(scenario)
}
