% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarScenario.R
\name{solarScenario_spec}
\alias{solarScenario_spec}
\title{Specification of a solar scenario}
\usage{
solarScenario_spec(
  model,
  from = "2010-01-01",
  to = "2010-12-31",
  theta = 0,
  exclude_known = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{model}{object with the class \code{solarModel}. See the function \code{\link{solarModel}} for details.}

\item{from}{character, start Date for simulations in the format \code{YYYY-MM-DD}.}

\item{to}{character, end Date for simulations in the format \code{YYYY-MM-DD}.}

\item{theta}{numeric, shift parameter for the mixture.}

\item{exclude_known}{when true the two starting points (equals for all the simulations) will be excluded from the output.}

\item{quiet}{logical}
}
\description{
Specification of a solar scenario
}
\note{
Version 1.0.0.
}
\examples{
model <- solarModel$new(spec)
model$fit()
simSpec <- solarScenario_spec(model)

}
\keyword{solarScenario}
