% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarScenario.R
\name{solarScenario_VaR}
\alias{solarScenario_VaR}
\title{Compute the Value at Risk from simulated values}
\usage{
solarScenario_VaR(scenarios, alpha = 0.05)
}
\arguments{
\item{scenarios}{An object of the class \code{solarScenario}}

\item{alpha}{Confidence level for the VaR}
}
\description{
Compute the Value at Risk from simulated values
}
\note{
Version 1.0.0.
}
\examples{
model <- Bologna
scen <- solarScenario(model, "2016-01-01", "2017-01-01", nsim = 10, by = "1 month")
solarScenario_VaR(scen, 0.05)
}
\keyword{solarScenario}
