% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarTransform-R6.R
\name{solarTransform}
\alias{solarTransform}
\title{Solar Model transformation functions}
\description{
Solar Model transformation functions

Solar Model transformation functions
}
\note{
Version 1.0.0.
}
\examples{
st <- solarTransform$new()

}
\keyword{solarModel}
\section{Super class}{
\code{\link[solarr:boundTransform]{solarr::boundTransform}} -> \code{solarTransform}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{epsilon}}{Numeric, \eqn{\epsilon} transformation parameter.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-solarTransform-new}{\code{solarTransform$new()}}
\item \href{#method-solarTransform-X}{\code{solarTransform$X()}}
\item \href{#method-solarTransform-iX}{\code{solarTransform$iX()}}
\item \href{#method-solarTransform-eta}{\code{solarTransform$eta()}}
\item \href{#method-solarTransform-ieta}{\code{solarTransform$ieta()}}
\item \href{#method-solarTransform-RY}{\code{solarTransform$RY()}}
\item \href{#method-solarTransform-iRY}{\code{solarTransform$iRY()}}
\item \href{#method-solarTransform-clone}{\code{solarTransform$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="solarr" data-topic="boundTransform" data-id="X_prime"><a href='../../solarr/html/boundTransform.html#method-boundTransform-X_prime'><code>solarr::boundTransform$X_prime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="boundTransform" data-id="Y"><a href='../../solarr/html/boundTransform.html#method-boundTransform-Y'><code>solarr::boundTransform$Y()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="boundTransform" data-id="bounds"><a href='../../solarr/html/boundTransform.html#method-boundTransform-bounds'><code>solarr::boundTransform$bounds()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="boundTransform" data-id="fit"><a href='../../solarr/html/boundTransform.html#method-boundTransform-fit'><code>solarr::boundTransform$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="boundTransform" data-id="g"><a href='../../solarr/html/boundTransform.html#method-boundTransform-g'><code>solarr::boundTransform$g()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="boundTransform" data-id="g_prime"><a href='../../solarr/html/boundTransform.html#method-boundTransform-g_prime'><code>solarr::boundTransform$g_prime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="boundTransform" data-id="iX_prime"><a href='../../solarr/html/boundTransform.html#method-boundTransform-iX_prime'><code>solarr::boundTransform$iX_prime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="boundTransform" data-id="iY"><a href='../../solarr/html/boundTransform.html#method-boundTransform-iY'><code>solarr::boundTransform$iY()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="boundTransform" data-id="ig"><a href='../../solarr/html/boundTransform.html#method-boundTransform-ig'><code>solarr::boundTransform$ig()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="boundTransform" data-id="print"><a href='../../solarr/html/boundTransform.html#method-boundTransform-print'><code>solarr::boundTransform$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="boundTransform" data-id="set_transform"><a href='../../solarr/html/boundTransform.html#method-boundTransform-set_transform'><code>solarr::boundTransform$set_transform()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="boundTransform" data-id="update"><a href='../../solarr/html/boundTransform.html#method-boundTransform-update'><code>solarr::boundTransform$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarTransform-new"></a>}}
\if{latex}{\out{\hypertarget{method-solarTransform-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{solarTransform} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$new(alpha = 0, beta = 1, link = "invgumbel")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{Numeric, \eqn{\alpha} transformation parameter.}

\item{\code{beta}}{Numeric, \eqn{\beta} transformation parameter.}

\item{\code{link}}{Character, link function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarTransform-X"></a>}}
\if{latex}{\out{\hypertarget{method-solarTransform-X}{}}}
\subsection{Method \code{X()}}{
Map the solar radiation \eqn{R_t} in the risk driver \eqn{X_t}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$X(Rt, Ct)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Rt}}{Numeric, solar radiation \eqn{R_t \in [C_{t}(1-\alpha-\beta), C_{t}(1-\alpha)]}.}

\item{\code{Ct}}{Numeric, clear sky radiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{\text{X}(R_t) = 1 - R_t/C_t}
}

\subsection{Returns}{
Numeric, risk driver \eqn{X_t \in (\alpha, \alpha+\beta)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarTransform-iX"></a>}}
\if{latex}{\out{\hypertarget{method-solarTransform-iX}{}}}
\subsection{Method \code{iX()}}{
Map the risk driver \eqn{X_t} in solar radiation \eqn{R_t}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$iX(Xt, Ct)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Xt}}{Numeric, risk driver in \eqn{ X_t \in (\alpha, \alpha+\beta)}.}

\item{\code{Ct}}{Numeric, clear sky radiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{\text{X}^{-1}(X_t) = C_t(1 - X_t)}
}

\subsection{Returns}{
Numeric, solar radiation \eqn{R_t \in [C_{t}(1-\alpha-\beta), C_{t}(1-\alpha)]}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarTransform-eta"></a>}}
\if{latex}{\out{\hypertarget{method-solarTransform-eta}{}}}
\subsection{Method \code{eta()}}{
Map the solar radiation \eqn{R_t} in the normalized variable \eqn{X_t^{\prime}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$eta(Rt, Ct)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Rt}}{Numeric, solar radiation \eqn{R_t \in [C_{t}(1-\alpha-\beta), C_{t}(1-\alpha)]}.}

\item{\code{Ct}}{Numeric, clear sky radiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{\eta(R_t) = \frac{1}{\beta}(1 - \alpha - R_t/C_t)}
}

\subsection{Returns}{
Numeric, normalized risk driver \eqn{X_t^{\prime} \in (0, 1)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarTransform-ieta"></a>}}
\if{latex}{\out{\hypertarget{method-solarTransform-ieta}{}}}
\subsection{Method \code{ieta()}}{
Map the normalized variable \eqn{X_t^{\prime}} to the solar radiatio \eqn{R_t}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$ieta(Xt_prime, Ct)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Xt_prime}}{Numeric, normalized risk driver \eqn{X_t^{\prime} \in (0, 1)}.}

\item{\code{Ct}}{Numeric, clear sky radiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{\eta^{-1}(X_t^{\prime}) = C_t(1 - \alpha - \beta \cdot X_t^{\prime})}
}

\subsection{Returns}{
Numeric, solar radiation \eqn{R_t \in [C_{t}(1-\alpha-\beta), C_{t}(1-\alpha)]}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarTransform-RY"></a>}}
\if{latex}{\out{\hypertarget{method-solarTransform-RY}{}}}
\subsection{Method \code{RY()}}{
Convert solar radiation \eqn{R_t} into the transformed variable \eqn{Y_t}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$RY(Rt, Ct)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Rt}}{Numeric, solar radiation \eqn{R_t \in [C_{t}(1-\alpha-\beta), C_{t}(1-\alpha)]}.}

\item{\code{Ct}}{Numeric, clear sky radiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{\text{RY}(R_t) = g\left(\frac{1}{\beta}(1 - \alpha- R_t/C_t)\right)}
}

\subsection{Returns}{
Transformed variable \eqn{Y_t \in (-\infty, \infty)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarTransform-iRY"></a>}}
\if{latex}{\out{\hypertarget{method-solarTransform-iRY}{}}}
\subsection{Method \code{iRY()}}{
Convert the transformed variable \eqn{Y_t} into solar radiation \eqn{R_t}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$iRY(Yt, Ct)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Yt}}{Numeric, transformed variable \eqn{Y_t \in (-\infty, \infty)}.}

\item{\code{Ct}}{Numeric, clear sky radiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{\text{iRY}(Y_t) = C_t(1 - \alpha-\beta g^{-1}(Y_t))}
}

\subsection{Returns}{
Numeric, solar radiation \eqn{R_t \in [C_{t}(1-\alpha-\beta), C_{t}(1-\alpha)]}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarTransform-clone"></a>}}
\if{latex}{\out{\hypertarget{method-solarTransform-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
