% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption-R6.R
\name{SoRadPorfolio}
\alias{SoRadPorfolio}
\title{Create a Solar Option Index portfolio}
\usage{
SoRadPorfolio(model, t_now, t_init, t_hor)
}
\description{
Create a Solar Option Index portfolio
}
\note{
Version 1.0.0.
}
\keyword{solarOption}
