% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_cor.R
\name{sp_cor_aniso}
\alias{sp_cor_aniso}
\alias{sp_cor_aniso_exp}
\alias{sp_cor_aniso_sph}
\title{Exponential and Spherical anisotropic spatial correlation}
\usage{
sp_cor_aniso_exp(h1, h2, phi)

sp_cor_aniso_sph(h1, h2, phi)
}
\arguments{
\item{h1}{Matrix of vector of distances.}

\item{h2}{Matrix of vector of distances.}

\item{phi}{Numeric vector, parameters.}
}
\description{
Exponential and Spherical anisotropic spatial correlation
}
\details{
Version 1.0.0. Implement the functions:
\deqn{\rho_{\exp}(h_1, h_2, \phi^{\exp}) = \exp\left(\frac{1}{\phi_1^{\exp}} \sqrt{\phi_2^{\exp} h_1^2 + h_2^2}\right)}
\deqn{\rho_{\text{sph}}(h_1, h_2, \phi^{\text{sph}}) =  1 - \frac{3}{2}\frac{\sqrt{\phi_2^{\text{sph}} h_1^2 + h_2}}{\phi_1^{\text{sph}}} + \frac{1}{2}\left(\frac{\sqrt{\phi_2^{\text{sph}} h_1^2 + h_2}}{\phi_1^{\text{sph}}}\right)^3}
}
