% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_cor.R
\name{sp_cor_isotr}
\alias{sp_cor_isotr}
\alias{sp_cor_isotr_exp}
\alias{sp_cor_isotr_gau}
\alias{sp_cor_isotr_sph}
\title{Exponential, Gaussian and Spherical isotropic spatial correlation}
\usage{
sp_cor_isotr_exp(h, phi)

sp_cor_isotr_gau(h, phi)

sp_cor_isotr_sph(h, phi)
}
\arguments{
\item{h}{Matrix of vector of distances.}

\item{phi}{Numeric scalar, parameter.}
}
\description{
Exponential, Gaussian and Spherical isotropic spatial correlation
}
\details{
Version 1.0.0. Implement the functions:
\deqn{\rho_{\exp}(h, \phi^{\exp}) = \exp\left(\frac{h}{\phi^{\exp}}\right)}
\deqn{\rho_{\text{gau}}(h, \phi^{\exp}) = \exp\left\{\left(\frac{h}{\phi^{\text{gau}}}\right)^2 \right\}}
\deqn{\rho_{\text{sph}}(h, \phi^{\text{sph}}) =  1 - \frac{3}{2}\frac{h}{\phi^{\text{sph}}} + \frac{1}{2}\left(\frac{h}{\phi^{\text{sph}}}\right)^3}
}
