% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGrid.R
\name{spatialGrid}
\alias{spatialGrid}
\title{Spatial Grid}
\value{
a tibble with two columns \code{lat} and \code{lon}.
}
\description{
Spatial Grid

Spatial Grid
}
\details{
Create a grid from a range of latitudes and longitudes.
}
\examples{
# Initialize a spatial grid
sp <- spatialGrid$new()

# Create an equally spaced grid
sp$make_grid(c(43.1, 44), c(9.2, 12.3), c(0.1, 0.1))
sp
# Check known point
sp$is_known_point(49.95, 12.15)
sp$is_known_point(43.9, 12)

# Check if a point is inside the bounds
sp$is_inside_bounds(44.8, 10.9)

# Extract a point
sp$neighborhoods(43.9, 12.1)

# Extract its neighborhoods
sp$neighborhoods(43.95, 12.15)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{weights}}{Weighting function for the distance.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{grid}}{A tibble with the spatial grid.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-spatialGrid-new}{\code{spatialGrid$new()}}
\item \href{#method-spatialGrid-set_grid}{\code{spatialGrid$set_grid()}}
\item \href{#method-spatialGrid-make_grid}{\code{spatialGrid$make_grid()}}
\item \href{#method-spatialGrid-is_inside_bounds}{\code{spatialGrid$is_inside_bounds()}}
\item \href{#method-spatialGrid-is_known_point}{\code{spatialGrid$is_known_point()}}
\item \href{#method-spatialGrid-known_point}{\code{spatialGrid$known_point()}}
\item \href{#method-spatialGrid-neighborhoods}{\code{spatialGrid$neighborhoods()}}
\item \href{#method-spatialGrid-print}{\code{spatialGrid$print()}}
\item \href{#method-spatialGrid-clone}{\code{spatialGrid$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialGrid-new"></a>}}
\if{latex}{\out{\hypertarget{method-spatialGrid-new}{}}}
\subsection{Method \code{new()}}{
Initialize a spatial grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialGrid$new(weights = IDW(2))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weights}}{Weighting function for the distance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialGrid-set_grid"></a>}}
\if{latex}{\out{\hypertarget{method-spatialGrid-set_grid}{}}}
\subsection{Method \code{set_grid()}}{
Set a spatial grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialGrid$set_grid(grid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{grid}}{Tibble with column \code{id}, \code{lat} and \code{lon}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialGrid-make_grid"></a>}}
\if{latex}{\out{\hypertarget{method-spatialGrid-make_grid}{}}}
\subsection{Method \code{make_grid()}}{
Create a spatial grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialGrid$make_grid(lat, lon, by, digits = 5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat}}{Numeric vector, from which is extracted the minimum and maximum for latitude.}

\item{\code{lon}}{Numeric vector, from which is extracted the minimum and maximum for longitude.}

\item{\code{by}}{Numeric vector, the first element is used to establish the distance between two latitudes in the grid.
The second element (if present) is used to establish the distance between two longitudes in the grid.}

\item{\code{digits}}{Integer scalar, number of digits for latitudes and longitudes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialGrid-is_inside_bounds"></a>}}
\if{latex}{\out{\hypertarget{method-spatialGrid-is_inside_bounds}{}}}
\subsection{Method \code{is_inside_bounds()}}{
Check if a location is inside the bounds of the grid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialGrid$is_inside_bounds(lat, lon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat}}{Numeric vector, reference latitudes.}

\item{\code{lon}}{Numeric vector, reference longitudes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} when the point is inside the limits and \code{FALSE} otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialGrid-is_known_point"></a>}}
\if{latex}{\out{\hypertarget{method-spatialGrid-is_known_point}{}}}
\subsection{Method \code{is_known_point()}}{
Check if a location is a known point inside the grid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialGrid$is_known_point(lat, lon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat}}{Numeric vector, reference latitudes.}

\item{\code{lon}}{Numeric vector, reference longitudes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} when the location is known and \code{FALSE} otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialGrid-known_point"></a>}}
\if{latex}{\out{\hypertarget{method-spatialGrid-known_point}{}}}
\subsection{Method \code{known_point()}}{
Return the id and coordinates of a location inside the grid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialGrid$known_point(lat, lon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat}}{Numeric vector, reference latitudes.}

\item{\code{lon}}{Numeric vector, reference longitudes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialGrid-neighborhoods"></a>}}
\if{latex}{\out{\hypertarget{method-spatialGrid-neighborhoods}{}}}
\subsection{Method \code{neighborhoods()}}{
Find the n-closest neighborhoods of a location.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialGrid$neighborhoods(lat, lon, n = 4)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat}}{Numeric scalar, reference latitude.}

\item{\code{lon}}{Numeric scalar, reference longitude.}

\item{\code{n}}{number of neighborhoods}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialGrid-print"></a>}}
\if{latex}{\out{\hypertarget{method-spatialGrid-print}{}}}
\subsection{Method \code{print()}}{
Method print for a \code{spatialGrid} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialGrid$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialGrid-clone"></a>}}
\if{latex}{\out{\hypertarget{method-spatialGrid-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialGrid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
