% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialKringing.R
\name{spatialKringing}
\alias{spatialKringing}
\title{\code{spatialKringing} object}
\description{
\code{spatialKringing} object

\code{spatialKringing} object
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{quiet}}{Logical}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{models}}{list of \code{kernelRegression} objects}

\item{\code{data}}{dataset with the parameters used for fitting}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-spatialKringing-new}{\code{spatialKringing$new()}}
\item \href{#method-spatialKringing-fit}{\code{spatialKringing$fit()}}
\item \href{#method-spatialKringing-predict}{\code{spatialKringing$predict()}}
\item \href{#method-spatialKringing-clone}{\code{spatialKringing$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialKringing-new"></a>}}
\if{latex}{\out{\hypertarget{method-spatialKringing-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{spatialKringing} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialKringing$new(data, params_names, grid, vg_models, quiet = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{dataset with spatial parameters and lon, lat.}

\item{\code{params_names}}{Names of the parameters to fit.}

\item{\code{grid}}{description}

\item{\code{vg_models}}{an optional list of kernelRegression models already fitted.}

\item{\code{quiet}}{description}

\item{\code{sample}}{List of parameter used as sample.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialKringing-fit"></a>}}
\if{latex}{\out{\hypertarget{method-spatialKringing-fit}{}}}
\subsection{Method \code{fit()}}{
Fit a \code{kernelRegression} object for a parameter or a group of parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialKringing$fit(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{list of parameters names to fit. When missing all the parameters will be fitted.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialKringing-predict"></a>}}
\if{latex}{\out{\hypertarget{method-spatialKringing-predict}{}}}
\subsection{Method \code{predict()}}{
Predict all the parameters for a specified location.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialKringing$predict(lat, lon, n = 4)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat}}{Numeric vector, latitudes in degrees.}

\item{\code{lon}}{Numeric vector, longitudes in degrees.}

\item{\code{n}}{Integer, number of neighborhoods to consider for interpolation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialKringing-clone"></a>}}
\if{latex}{\out{\hypertarget{method-spatialKringing-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialKringing$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
