% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialModel.R
\name{spatialModel}
\alias{spatialModel}
\title{Spatial model object}
\description{
Spatial model object

Spatial model object
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{quiet}}{logical, when \code{TRUE} the function will not display any message.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{grid}}{object with the spatial grid}

\item{\code{models}}{list of \code{solarModel} objects}

\item{\code{parameters}}{\code{spatialParameters} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-spatialModel-new}{\code{spatialModel$new()}}
\item \href{#method-spatialModel-get}{\code{spatialModel$get()}}
\item \href{#method-spatialModel-interpolate}{\code{spatialModel$interpolate()}}
\item \href{#method-spatialModel-solarModel}{\code{spatialModel$solarModel()}}
\item \href{#method-spatialModel-clone}{\code{spatialModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-spatialModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize the spatial model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialModel$new(models, paramsModels, quiet = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{models}}{A list of \code{solarModel} objects}

\item{\code{paramsModels}}{A \code{spatialParameters} object.}

\item{\code{quiet}}{logical}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialModel-get"></a>}}
\if{latex}{\out{\hypertarget{method-spatialModel-get}{}}}
\subsection{Method \code{get()}}{
Get a known model in the grid from place or coordinates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialModel$get(id, lat, lon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{character, id of the location.}

\item{\code{lat}}{numeric, latitude of a location.}

\item{\code{lon}}{numeric, longitude of a location.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialModel-interpolate"></a>}}
\if{latex}{\out{\hypertarget{method-spatialModel-interpolate}{}}}
\subsection{Method \code{interpolate()}}{
Perform the bilinear interpolation for a target variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialModel$interpolate(lat, lon, target = "GHI", n = 4, day_date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat}}{numeric, latitude of the location to be interpolated.}

\item{\code{lon}}{numeric, longitude of the location to be interpolated.}

\item{\code{target}}{character, name of the target variable to interpolate.}

\item{\code{n}}{number of neighborhoods to use for interpolation.}

\item{\code{day_date}}{date for interpolation, if missing all the available dates will be used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialModel-solarModel"></a>}}
\if{latex}{\out{\hypertarget{method-spatialModel-solarModel}{}}}
\subsection{Method \code{solarModel()}}{
Interpolator function for a \code{solarModel} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialModel$solarModel(lat, lon, n = 4)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat}}{numeric, latitude of a point in the grid.}

\item{\code{lon}}{numeric, longitude of a point in the grid.}

\item{\code{n}}{number of neighborhoods}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-spatialModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
