% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialScenario.R
\name{spatialScenario_residuals}
\alias{spatialScenario_residuals}
\title{Simulate residuals from a a \code{spatialScenario_spec}}
\usage{
spatialScenario_residuals(simSpec, nsim = 1, seed = 1)
}
\arguments{
\item{simSpec}{object with the class \code{spatialScenario_spec}. See the function \code{\link{spatialScenario_spec}} for details.}

\item{nsim}{integer, number of simulations.}

\item{seed}{scalar integer, starting random seed.}
}
\description{
Simulate residuals from a a \code{spatialScenario_spec}
}
