% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialScenario.R
\name{spatialScenario_spec}
\alias{spatialScenario_spec}
\title{Specification of a solar scenario}
\usage{
spatialScenario_spec(
  sm,
  sc,
  places,
  from = "2010-01-01",
  to = "2010-01-31",
  exclude_known = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{sm}{\code{spatialModel} object}

\item{sc}{\code{spatialCorrelation} object}

\item{places}{target places}

\item{from}{character, start Date for simulations in the format \code{YYYY-MM-DD}.}

\item{to}{character, end Date for simulations in the format \code{YYYY-MM-DD}.}

\item{exclude_known}{when true the two starting points (equals for all the simulations) will be excluded from the output.}

\item{quiet}{logical}
}
\description{
Specification of a solar scenario
}
