% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralDistribution.R
\name{spectralDistribution}
\alias{spectralDistribution}
\title{Compute the spectral distribution for a black body}
\usage{
spectralDistribution(x, measure = "nanometer")
}
\arguments{
\item{measure}{character, measure of the irradiated energy. If \code{nanometer} the final energy will be in W/m2 x nanometer,
otherwise if \code{micrometer} the final energy will be in W/m2 x micrometer.}

\item{lambda}{numeric, wave length in micrometers.}
}
\description{
Compute the spectral distribution for a black body
}
