% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsugeno.R
\name{sugeno_bounds}
\alias{sugeno_bounds}
\title{Sugeno upper and lower parameters.}
\usage{
sugeno_bounds(lambda)
}
\arguments{
\item{lambda}{Numeric, distortion parameter.}
}
\description{
Sugeno upper and lower parameters.
}
